% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{circle_text_func}
\alias{circle_text_func}
\title{circle_text_func}
\usage{
circle_text_func(
  g,
  circ_layout,
  text.label = NULL,
  text.size = 0.8,
  text.distance = 1.5,
  text.colour = "black",
  ...
)
}
\arguments{
\item{g}{igraph object. Input network object.}

\item{circ_layout}{layout object. The circular layout object for \code{g}, usually derived from function \code{\link{igraph::layout.circle()}}.}

\item{text.label}{string vector. The label to display. Default is \code{NULL}.}

\item{text.size}{numeric or numeric vector. The text size for the labels. Default is \code{0.8}.}

\item{text.distance}{numeric. The distance multiplier between label and nodes. Default is \code{1.5}.}

\item{text.colour}{string. Label colour. Default is \code{"black"}.}
}
\value{
Added text on the igraph plot.
}
\description{
Companion function for the network visualization function: used for displaying node labels around
             the circular network figures.
}
\details{
The \code{text.size} argument also accepts a vector of sizes with a length equal to the number of vertices.
        When different unequal length is detect, the function uses the first number for a universal text size.
}
\examples{
\dontrun{
     # plot the original circular plot
     g_layout <- layout.circle(g)
     plot(
       g,
       layout=g_layout,
       edge.curved=TRUE,
       vertex.size=vSizes,
       vertex.label = NA,
       vertex.label.dist=-1.5,
       vertex.label.color="black",
       vertex.label.cex=0.8,
       asp=FALSE,
       vertex.label.cex=0.6,
       edge.width=edgeweights,
       edge.arrow.mode=0,
       main="")

    # add labels to the plot
     circle_text_func(g = g, circ_layout = g_layout)
}
}
