% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_pvalue}
\alias{cor_pvalue}
\title{cor_pvalue}
\usage{
cor_pvalue(r, n)
}
\arguments{
\item{r}{Correlation coefficient}

\item{n}{Sample size}
}
\value{
P value for the correlation analysis
}
\description{
Calculate p value for the correlation analysis
}
\examples{
\dontrun{

p <- cor_pvalue(r = 0.806687547, n = 6)

}
}
