% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{GSAscatterHT}
\alias{GSAscatterHT}
\title{rbioArray_allGSscatter}
\usage{
GSAscatterHT(GSAList, rankCutoff, pCutoff, fileName, plotWidth = 170,
  plotHeight = 150)
}
\arguments{
\item{GSAList}{GSA list generated from \code{\link{rbioArray_allGSA}}.}

\item{rankCutoff}{Cutoff value for GS rank line.}

\item{pCutoff}{Cutoff value for GS p value line.}

\item{fileName}{Output file name.}

\item{plotHeight}{Set the height of the plot. Default is \code{150}.}

\item{GS}{if GS = \code{"KEGG"}, the function will remove the "KEGG_" string in the GS name variable. Default is \code{"OTHER"}.}
}
\value{
Outputs a \code{pdf} scatter figure file with allGSA results.
}
\description{
Generate scatter plot for piano GS rank heatmap obejct.
}
\details{
The function is based on piano package.
}
\examples{
\dontrun{
pcGObp_Htmap <- consensusHeatmap(pc_GObp, cutoff = 15,method="median",
 colorkey = TRUE, colorgrad = c("blue","white"),
 cellnote = "none")
}
}

