% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{microarray_de}
\alias{microarray_de}
\title{microarray_de}
\usage{
microarray_de(object, contra)
}
\arguments{
\item{object}{The input \code{rbioarray_flist} object.}

\item{contra}{contra Contrast matrix.}
}
\value{
A \code{rbioarray_de} class object with DE results, containing the following core items:

        \code{F_stats}

        \code{DE_results}: a list containing the DE results

        \code{comparisons}

        \code{fit}: the limma fitted DE object as a reference

        Additionally, the \code{rbioarray_de} includes additional items from the input \code{rbioarray_flist} object.
}
\description{
Function that performs statistical analysis for microarray data from \code{rbioarray_flist} class object.
}
\examples{
\dontrun{

de_list <- microarray_de(object = filtered_list, contra = contra)

}
}
