% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbio_entrez2geneStats}
\alias{rbio_entrez2geneStats}
\title{rbioArray_entrez2gene}
\usage{
rbio_entrez2geneStats(DEGdfm, cat = "SYMBOL", species = "Hs",
  pkg = "org.Hs.eg.db")
}
\arguments{
\item{DEGdfm}{Input DE data frame.}

\item{cat}{Catergory, default is \code{"SYMBOL"}.}

\item{species}{Set the species, default is \code{"Hs"}.}

\item{pkg}{Name for the annotation package. The current package includes the human version.}
}
\value{
Outputs a \code{dataframe} object with Entrez ID.
}
\description{
Add Human entrez ID to the DE dataframe
}
\examples{
\dontrun{
DE_dataframe <- rbio_entrez2gene(dataframe)
}
}

