% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{rbio_kmeans}
\alias{rbio_kmeans}
\title{rbio_kmeans}
\usage{
rbio_kmeans(
  x,
  export.name = NULL,
  k_range = 2:(nrow(unique(x)) - 1),
  nstart = 25,
  plot.title = NULL,
  plot.titleSize = 10,
  plot.symbolSize = 2,
  plot.vline.label = FALSE,
  plot.fontType = "sans",
  plot.xLabel = "Number of clusters (K)",
  plot.xLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.xTickItalic = FALSE,
  plot.xTickBold = FALSE,
  plot.xAngle = 0,
  plot.xhAlign = 0.5,
  plot.xvAlign = 0.5,
  plot.yLabel = "Average silhouette score",
  plot.yLabelSize = 10,
  plot.yTickLblSize = 10,
  plot.yTickItalic = FALSE,
  plot.yTickBold = FALSE,
  plot.rightsideY = TRUE,
  plot.width = 170,
  plot.height = 150,
  verbose = TRUE
)
}
\arguments{
\item{x}{Matrix or Data frame. Input data matrix. The function will cluster row items.}

\item{export.name}{Optional user defined export name prefix. Default is \code{NULL}.}

\item{k_range}{Integer array. The range of K to try to find the optimal number of cluster (K). Default is \code{2:15}.}

\item{nstart}{The number of tries to find the optimal cluster per K. Default is \code{25}.}

\item{plot.title}{Whether to display plot title on top of the plot. Default is \code{FALSE}.}

\item{plot.titleSize}{The font size of the plot title. Default is \code{10}.}

\item{plot.symbolSize}{Symbol size. Default is \code{2}.}

\item{plot.vline.label}{Boolean. If to show the label the on the silhouette}

\item{plot.fontType}{The type of font in the figure. Default is "sans". For all options please refer to R font table, which is available on the website: \url{http://kenstoreylab.com/?page_id=2448}.}

\item{plot.xLabel}{x axis label. Type with quotation marks. Could be NULL. Default is \code{"Features"}.}

\item{plot.xLabelSize}{x axis label size. Default is \code{10}.}

\item{plot.xTickLblSize}{Font size of x axis ticks. Default is \code{10}.}

\item{plot.xTickItalic}{Set x axis tick font to italic. Default is \code{FALSE}.}

\item{plot.xTickBold}{Set x axis tick font to bold. Default is \code{FALSE}.}

\item{plot.xAngle}{The rotation angle (degrees) of the x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.yLabel}{y axis label. Type with quotation marks. Could be NULL. Default is \code{"OOB error rate"}.}

\item{plot.yLabelSize}{y axis label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Font size of y axis ticks. Default is \code{10}.}

\item{plot.yTickItalic}{Set y axis tick font to italic. Default is \code{FALSE}.}

\item{plot.yTickBold}{Set y axis tick font to bold. Default is \code{FALSE}.}

\item{plot.rightsideY}{If to display the right side y-axis. Default is \code{TRUE}.}

\item{plot.width}{The width of the plot (unit: mm). Default is 170. Default will fit most of the cases.}

\item{plot.height}{The height of the plot (unit: mm). Default is 150. Default will fit most of the cases.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
A silhouette score plot showing the optimized number of clusters (K), and a \code{kmeans} object containing the final K means cluster result.
}
\description{
K means cluster function
}
\details{
The function does three things:
         1. Conducts a silhouette score test to find out the optimal K (number of clusters)
         2. Exports the silhouette score plot
         3. Conducts the K means cluster analysis with the optimized K

         For the input \code{x}, it typically does not contain annotation columns for microarray or RNA-seq data sets.
         The function clusters the row items.
         To have proper labels for the cluser plot using \code{rbio_kmeans_plot()}, make sure to have rownames set for \code{x}.
}
\examples{
\dontrun{
mtx <- group_de_list$ex1$E
rownames(mtx) <- group_de_list$ex1$genes$SYMBOL
rbio_kmeans(x = mtx, export.name = "ex1", plot.vline.label = TRUE)
}
}
