% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{rbio_kmeans_plot}
\alias{rbio_kmeans_plot}
\title{rbio_kmeans_plot}
\usage{
rbio_kmeans_plot(
  km.object,
  x,
  export.name = NULL,
  centerScale = TRUE,
  ...,
  plot.comps = c(1:2),
  plot.Title = NULL,
  plot.sampleLabel.type = "none",
  plot.sampleLabelSize = 2,
  plot.sampleLabel.padding = 0.5,
  plot.SymbolSize = 2,
  plot.ellipse = FALSE,
  plot.ellipse_conf = 0.95,
  plot.xLabelSize = 10,
  plot.yLabelSize = 10,
  plot.xTickLblSize = 10,
  plot.yTickLblSize = 10,
  plot.xAngle = 0,
  plot.xhAlign = 0.5,
  plot.xvAlign = 0.5,
  plot.mtx.densityplot = FALSE,
  plot.mtx.stripLblSize = 10,
  plot.rightsideY = FALSE,
  plot.fontType = "sans",
  plot.Width = 170,
  plot.Height = 150,
  verbose = TRUE
)
}
\arguments{
\item{km.object}{A \code{kmeans} object, from \code{\link{rbio_kmeans()}} function.}

\item{x}{Matrix or data.frame. The same data used to generate the input \code{km.object}.}

\item{centerScale}{Boolean. If to center and scale the x data fro clustering. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{prcomp} function.}

\item{plot.comps}{Integer or vector of integers. Index number(s) for principal component(s) to plot. Default is \code{c(1, 2)}.}

\item{plot.Title}{The biplot title. Default is \code{NULL}.}

\item{plot.sampleLabel.type}{If to show the sample labels on the graph. Options are \code{"none"}, \code{"direct"} and \code{"indirect"}. Default is \code{"none"}.}

\item{plot.sampleLabelSize}{Only set when \code{biplot.sampleLabel.type} is not \code{"none"}, The size of the sample label. Default is \code{2}.}

\item{plot.sampleLabel.padding}{Set only when \code{biplot.sampleLabel.type = "indirect"}, the padding between sample symbol and the label. Default is \code{0.5}.}

\item{plot.SymbolSize}{The symbol size for the scatter plot portion of the biplot. Default is \code{2}.}

\item{plot.ellipse}{If to draw ellipses. Default is \code{FALSE}.}

\item{plot.ellipse_conf}{The confidence value for the ellipses. Default is \code{0.93}.}

\item{plot.xTickLblSize}{X-axis tick label size. Default is \code{10}.}

\item{plot.yTickLblSize}{Y-axis tick label size. Default is \code{10}.}

\item{plot.xAngle}{The rotation angle (degrees) of the biplot x axis marks. Default is \code{0} - horizontal.}

\item{plot.xhAlign}{The horizontal alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.xvAlign}{The vertical alignment type of the x axis marks. Options are \code{0}, \code{0.5} and \code{1}. The default value at \code{0} is especially useful when \code{xAngle = 90}.}

\item{plot.mtx.densityplot}{If to display a density plot on the diagonal for the correlation scoreplot matrix. Default is \code{FALSE}.}

\item{plot.mtx.stripLblSize}{The label font size for the correlation scoreplot matrix strips. Default is \code{10}.}

\item{plot.rightsideY}{If to show the right side y-axis for both boxplot and biplot. For biplot, only applicable when the length of \code{comps} is less than 2, inclusive. Default is \code{FALSE}.}

\item{plot.fontType}{Font for the figure texts. Default is \code{"sans"}.}

\item{plot.Width}{The biplot width. Default is \code{170}.}

\item{plot.Height}{The biplot height. Default is \code{150}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{plot.loadingplot}{If to superimpose loading plot. Default is \code{FALSE}.}

\item{plot.loadingplot.textsize}{The font size of the loading plot labels. Default is \code{3}.}
}
\value{
Outputs a PCA object, a boxplot (proportion of variance) and a biplot from PCA analysis. The format is \code{pdf}.
}
\description{
A plot function for K means cluster results, with or without PCA.
}
\details{
To have proper labels for the cluser plot using \code{rbio_kmeans_plot()}, make sure to have rownames set for \code{x}.

         The function automatically detects if PCA is needed:
         when \code{x} has more than two variables, the function will run PCA first.
         \code{plot.comps} not needed and will not be used when \code{x} has less than three variables, i.e. no PCA needed for plotting.
}
\examples{
\dontrun{
rbio_kmeans_plot(km.object = ex1_kmeans, x = mtx, export.name = "ex1",
                 plot.ellipse = TRUE,
                 plot.sampleLabel.type = "direct",
                 plot.SymbolSize = 4,
                 plot.xLabelSize = 14, plot.yLabelSize = 14,
                 plot.Width = 140, plot.Height = 150)
}
}
