% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{rbio_randomforest_fs}
\alias{rbio_randomforest_fs}
\title{rbio_randomforest_fs}
\usage{
rbio_randomforest_fs(
  object,
  sample_id.var = NULL,
  sample_group.var = NULL,
  export.name = NULL,
  gene_symbol.only = FALSE,
  fs_on_de = FALSE,
  rffs.center_scale = FALSE,
  rffs.quantile = FALSE,
  rffs.ntimes = 50,
  rffs.inital_fs.ntree = 501,
  rffs.inital_fs.mtry = NULL,
  rffs.sfs = FALSE,
  rffs.sfs_ntree = 501,
  rffs.randomstate = 0,
  parallelComputing = TRUE,
  n_cores = parallel::detectCores() - 1,
  clusterType = "FORK"
)
}
\arguments{
\item{object}{\code{sig} object. Input \code{sig} class object.}

\item{sample_id.var}{String. Sample id variable name, present in \code{object$input_data$targets}.}

\item{export.name}{String. Prefix for export files and objects. When \code{NULL}, the function uses the input object name.}

\item{gene_symbol.only}{Boolean. Whether or not to remove probes without gene symbol. Default is \code{FALSE}.}

\item{fs_on_de}{Boolean. Whether to conduct RFFS on only the DE targets (features), according to the iput object.}

\item{rffs.center_scale}{Boolean. Whether to center.scale the data for RFFS. Default is \code{FALSE}.}

\item{rffs.quantile}{Boolean. Whether to quantile normalize the data for RFFS. Default is \code{FALSE}.}

\item{rffs.ntimes}{Integer. Number of random forest recursion. Default is \code{50}.}

\item{rffs.inital_fs.ntree}{Integer. ntree setting for the RFFS initial selection. Default is \code{501}.}

\item{rffs.inital_fs.mtry}{Integer. mtry setting for the RFFS initial selection. Default is \code{NULL}.}

\item{rffs.sfs}{Boolean. Whether to run the secondary sequential forward selection (SFS). Default is \code{FALSE}.}

\item{rffs.sfs_ntree}{Integer. Set when \code{rffs.sfs = TRUE}, ntree setting for the RFFS SFS step. Default is \code{501}.}

\item{rffs.randomstate}{Integer. Set random state for RFFS. If \code{0}, the random state is off. Default is \code{0}.}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Set when \code{parallelComputing = TRUE}, number of CPU threads/cores for parallel computing.}

\item{sample_group.car}{String. Sample group variable name, present in \code{object$input_data$targets}.}

\item{cluterType}{clusterType Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Export an \code{rbio_rffs} object to the global environment.
}
\description{
Supervised feature selection with recursive random forest (RFFS)
}
\details{
The resulting \code{rbio_rffs} object contains the following items:

         \code{ml_raw_dataframe} Raw data dataframe for RFFS. Data not normalized or center.scaled by the function.
                                   The data is usually previously normalized or processed.
                                   The E data is from \code{object$input_data$norm_E} from the input \code{sig} object.
                                   It is noted the data may be subset with only the DE features according to the \code{fs_on_de} setting.
                                   This data frame can be used as input data for other ML purpouses featured in the \code{RBioFS} packages.

         \code{rffs_working_data} It is list containing the follow items:
             \code{rffs_working_E} Working datafarme as the \code{x} input for RFFS. Columns: features; Rows: samples. Normalized/center.scaled when set.
             \code{rffs_working_y}  Working vector (\code{factor}) as the \code{y} input for RFFS.#'
             \code{center_scale} Center.scale setting.
             \code{quantile} Quantile setting.

         \code{rffs_initial_results}

         \code{rffs_sfs_results}

         \code{rffs_subset_dataframe}: Input data subset with only the RFFS selected features. No data or fas

         \code{rffs_ntimes}

         \code{rffs_inital_ntree}

         \code{rffs_inital_mtry}

         \code{rffs_sfs}

         \code{rffs_sfs_ntree}
}
