% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{rbio_supervised_hcluster}
\alias{rbio_supervised_hcluster}
\title{rbio_supervised_hcluster}
\usage{
rbio_supervised_hcluster(
  object,
  gene_symbol.only = FALSE,
  sample_id.var.name = NULL,
  distance = c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"),
  clust = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty", "median",
    "centroid"),
  col.colour = "Paired",
  map.colour = "PRGn",
  n.map.colour = 11,
  ...,
  plot.width = 7,
  plot.height = 7,
  verbose = TRUE
)
}
\arguments{
\item{object}{Input \code{sig} class object.}

\item{gene_symbol.only}{Whether or not to remove probes without gene symbol. Default is \code{FALSE}.}

\item{distance}{Distance calculation method. Default is \code{"euclidean"}. See \code{\link{dist}} for more.}

\item{clust}{Clustering method. Default is \code{"complete"}. See \code{\link{hclust}} for more.}

\item{col.colour}{Column group colour. Default is \code{"Paired"}. See \code{RColorBrewer} package for more.}

\item{map.colour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n.map.colour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{plot.width}{Width of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{plot.height}{Height of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{sample_id.vector}{A \code{vector} containing names to display for each heatmap column. Default is \code{NULL} and the function will use the column name from the input.}
}
\value{
A heatmap based on hierarchical clustering analysis in \code{pdf} format.
}
\description{
Supersived hierarchical clustering function.
}
\details{
Unlike the unsupervised veresion, the supervised hcluster uses normalized expression data for both RNAseq and microaray.
         The column colour group is usually 2.
}
