% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rbio_tom}
\alias{rbio_tom}
\title{rbio_tom}
\usage{
rbio_tom(
  mtx,
  diag = FALSE,
  cor_method = c("pearson", "kendall", "spearman"),
  power = 6,
  tom_type = c("unsigned", "signed"),
  ...,
  hclust.method = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty",
    "median", "centroid"),
  k = NULL,
  h = NULL,
  plot.dendro = TRUE,
  plot.export.name = NULL,
  plot.margins = c(0.5, 0.5, 0.5, 0.5),
  plot.title = NULL,
  plot.title.size = 16,
  plot.dendroline.size = 0.8,
  plot.dendrolabel = TRUE,
  plot.dendrolabel.size = 1.5,
  plot.dendrolabel.space = 2,
  plot.ylabel.size = 1.5,
  plot.width = 150,
  plot.height = 150,
  verbose = TRUE
)
}
\arguments{
\item{mtx}{matrix. TBC.}

\item{diag}{Boolean. If to include diagonal for igraph construction. Default is \code{FALSE}.}

\item{cor_method}{String. Correlation method. Default is \code{"pearson"}.}

\item{power}{integer. TBC.}

\item{tom_type}{string. TBC.}

\item{...}{Additional arguments passed to \code{TOMdist()} function.}

\item{hclust.method}{String.}

\item{k}{integer. TBC.}

\item{h}{numeric. TBC.}

\item{plot.dendro}{Boolean. Whether to plot a dendrogram. Default is \code{TRUE}.}

\item{plot.export.name}{string. The prefix for the exported figure file name. Default is \code{NULL}.}

\item{plot.margins}{numeric vector. Plot margins, unit is "cm", order: \code{t, r, b, l}. Default is \code{c(2, 2, 2, 2)}.}

\item{plot.title}{Boolean. The dendrogram plot title. Default is \code{NULL}.}

\item{plot.title.size}{numeric. The dendrogram plot title size. Default is \code{16}.}

\item{plot.dendroline.size}{numeric. TBC.}

\item{plot.dendrolabel}{Boolean. TBC.}

\item{plot.dendrolabel.size}{mumeric. TBC.}

\item{plot.dendrolabel.space}{numeric. TBC.}

\item{plot.ylabel.size}{numeric. TBC.}

\item{plot.width}{numeric. The dendrogram plot width. Default is \code{170}.}

\item{plot.height}{numeric. The dendrogram plot height. Default is \code{150}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
TBC
}
\description{
TOM (topological overlap measure) analysis.
}
\details{
TBC

        The \code{plot.margins} follow the base R setting in \code{\link{par}} for the positioning:
        b: mar[1], l: mar[2], t: mar[3], r: mar[4]
}
\examples{
\dontrun{
           tom_g <- rbio_tom(mtx = t(mtx), k = 10,
                             plot.export.name = "ex1",
                             plot.title = "TOM distance hcluster", plot.title.size = 16,
                             plot.margins = c(0.5 ,0.5, 0.5, 0.5),
                             plot.dendroline.size = 0.2,
                             plot.dendrolabel = TRUE,
                             plot.dendrolabel.size = 1.5,
                             plot.dendrolabel.space = 2,
                             plot.ylabel.size = 1.5)
}
}
