% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_unsupervised_corcluster.default}
\alias{rbio_unsupervised_corcluster.default}
\title{rbio_unsupervised_corcluster.default}
\usage{
\method{rbio_unsupervised_corcluster}{default}(E, genes,
  input.sample_groups, n = "all", input.comparisons = NULL,
  cor.method = "pearson", rm.control = FALSE,
  input.genes_annotation.control_type, gene_symbol.only = FALSE,
  input.genes_annotation.gene_symbol.var_name = NULL,
  input.genes_annotation.gene_id.var_name = NULL, map.colour = "PRGn",
  n.map.colour = 11, ..., export.name = NULL,
  heatmap.axis.label = FALSE, heatmap.width = 7, heatmap.height = 7,
  sigplot = FALSE, sigplot.adj.p = FALSE, sigplot.alpha = 0.05,
  sigplot.sigLabelColour = "red", sigplot.sigLabelSize = 3,
  sigplot.labelColour = "black", sigplot.labelSize = 1,
  sigplot.labelAngle = 90, sigplot.keySize = 1, sigplot.mar = c(5, 4,
  4, 2) + 0.1, sigplot.width = 7, sigplot.height = 7)
}
\arguments{
\item{E}{Expression or count matrix, with rows for genes/probes/genomic features, columns for RNA samples.}

\item{genes}{Annotation data frame for genes/probes/genomic features.}

\item{input.sample_groups}{Input \code{factor} object for sample groupping labels.}

\item{n}{Number of genes/probes/genomic features to be clustered, numeric input or \code{"all"}. Default is \code{"all"}.}

\item{cor.method}{The correlation method, options are "pearson", "spearman" and "pearson". Default is \code{"pearson"}.}

\item{rm.control}{Whether to remove control probes (Agilent platform) or not. Default is \code{TRUE}.}

\item{input.genes_annotation.control_type}{Only set when \code{rm.control = TRUE}, input control type variable annotation list.}

\item{gene_symbol.only}{Whether or not to remove probes without gene symbol. Default is \code{FALSE}.}

\item{input.genes_annotation.gene_symbol.var_name}{Only set when \code{gene_symbol.only = TRUE}, variable name for gene symbol column in \code{genes} data frame.}

\item{input.genes_annotation.gene_id.var_name}{Only set when \code{gene_symbol.only = TRUE}, variable name for gene id column in \code{genes} data frame.}

\item{map.colour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n.map.colour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{export.name}{File name for the export \code{pdf} plot file.}

\item{heatmap.axis.label}{Whether to display label for both x- and y- axes. Default is \code{FALSE}.}

\item{heatmap.width}{Width for correlation heatmap. Unit is \code{inch}. Default is \code{7}.}

\item{heatmap.height}{Height for correlation heatmap. Unit is \code{inch}. Default is \code{7}.}

\item{sigplot}{If to show significance plot. Default is \code{FALSE}.}

\item{sigplot.adj.p}{If to use FDR corrected correlation p value to plot sigPlot. Default is \code{FALSE}.}

\item{sigplot.alpha}{The alpha value for correlation p value. Default is \code{0.05}}

\item{sigplot.sigLabelColour}{The colour for label for the significant pairs. Default is \code{"red"}.}

\item{sigplot.sigLabelSize}{The size for label for the significant pairs. Default is \code{3}.}

\item{sigplot.labelColour}{The significance heatmap axis label colour. Default is \code{"black"}.}

\item{sigplot.labelSize}{The significance heatmap axis label size. Default is \code{1}.}

\item{sigplot.labelAngle}{The significance heatmap axis label angle. Default is \code{90}.}

\item{sigplot.keySize}{The significance heatmap colour key size. Default is \code{1}.}

\item{sigplot.mar}{The A numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot. The default is c(5, 4, 4, 2) + 0.1.}

\item{sigplot.width}{Width for correlation significance plot. Unit is \code{inch}. Default is \code{7}.}

\item{sigplot.height}{Height for correlation significance plot. Unit is \code{inch}. Default is \code{7}.}
}
\value{
A heatmap based on hierarchical clustering analysis in \code{pdf} format.
}
\description{
Default unsupersived correlation clustering function.
}
