% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_unsupervised_corcluster.rbioseq_de}
\alias{rbio_unsupervised_corcluster.rbioseq_de}
\title{rbio_unsupervised_corcluster.rbioseq_de}
\usage{
\method{rbio_unsupervised_corcluster}{rbioseq_de}(object, sample_id.var.name = NULL, ...)
}
\arguments{
\item{object}{Input object in \code{rbioseq_de} class.}

\item{...}{Additional arguments for the default method.}
}
\description{
\code{rbio_unsupervised_corcluster} function for \code{rbioseq_de} class object.
}
\details{
The function uses filtered count data, as opposed to normalized data.
         Due to the compositional nature of NGS data, the count data is transformed using CLR method prior to clustering.
}
