% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_unsupervised_corcluster}
\alias{rbio_unsupervised_corcluster}
\title{rbio_unsupervised_corcluster}
\usage{
rbio_unsupervised_corcluster(object, ...)
}
\arguments{
\item{object}{Input object in either \code{rbioseq_de} or \code{rbioarray_de} class.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
A pdf file containing a heatmap for unsupervised hierarchical clustering analysis.
}
\description{
Generic unsupersived correlation clustering function.
}
\details{
It behaves in the following ways:
         1. heatmap uses \code{heatmap.2} function from \code{gplots} package.
         2. significance plot uses \code{corrplot} package

         Since the function depends on the \code{heatmap.2} function from \code{gplots} package for heatmap,
         arguments can be passed directly, as seen in the examples below.
}
\examples{

\dontrun{
# rbioarray_de class input

# rbioseq_de class input

}

}
