% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{rbio_unsupervised_hcluster.rbioarray_flist}
\alias{rbio_unsupervised_hcluster.rbioarray_flist}
\title{rbio_unsupervised_hcluster.rbioarray_flist}
\usage{
\method{rbio_unsupervised_hcluster}{rbioarray_flist}(object,
  sample_id.var.name = NULL, ...)
}
\arguments{
\item{object}{Input object in \code{rbioarray_flist} class.}

\item{sample_id.var.name}{Variable name for sample identification, typically from \code{object$target}.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
A pdf file containing a heatmap for unsupervised hierarchical clustering analysis.
}
\description{
\code{rbio_unsupervised_hcluster} function for \code{rbioarray_flist} class object.
}
