% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{rbio_unsupervised_hcluster.rbioseq_de}
\alias{rbio_unsupervised_hcluster.rbioseq_de}
\title{rbio_unsupervised_hcluster.rbioseq_de}
\usage{
\method{rbio_unsupervised_hcluster}{rbioseq_de}(object,
  sample_id.var.name = NULL, ...)
}
\arguments{
\item{object}{Input object in \code{rbioseq_de} class.}

\item{sample_id.var.name}{Variable name for sample identification, typically from \code{object$target}.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
A pdf file containing a heatmap for unsupervised hierarchical clustering analysis.
}
\description{
\code{rbio_unsupervised_hcluster} function for \code{rbioseq_de} class object.
}
\details{
The function uses filtered count data, as opposed to normalized data.
         Due to the compositional nature of NGS data, the count data is transformed using CLR method prior to clustering.
}
