% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioArray_allGSA}
\alias{rbioArray_allGSA}
\title{rbioArray_allGSA}
\usage{
rbioArray_allGSA(MyGS, pVar, logFCVar, tVar)
}
\arguments{
\item{MyGS}{Pre-loaded gene set objects.}

\item{pVar}{Gene level p values. Could be, but not exclusive to, a variable of a dataframe.}

\item{logFCVar}{Gene level logFC (log fold change). Could be, but not exclusive to, a variable of a dataframe.}

\item{tVar}{Gene leve t values. Could be, but not exclusive to, a variable of a dataframe.#'}
}
\value{
Outputs a \code{list} object with GSA results, both for p value based and t value based.
}
\description{
Add Human entrez ID to the DE dataframe
}
\details{
The function is based on piano package. It runs "fisher", "stouffer", "reporter", "tailStrength", "wilcoxon" for p value based GSA, and "page", "gsea", "maxmean" for t value based GSA.
}
\examples{
\dontrun{
GOterm_bp <- loadGSC(file = "c5.bp.v5.0.entrez.gmt", type = "gmt") # load GO term biological process set

pc_Kegg <- rbioArray_allGSA(GS_Kegg, pVar = pcGSdfm$p_value, logFCVar = pcGSdfm$logFC, tVar = pcGSdfm$t_value)

}
}

