% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioArray_allGSbox}
\alias{rbioArray_allGSbox}
\title{rbioArray_allGSbox}
\usage{
rbioArray_allGSbox(GSrank, GS = "OTHER", fileName, plotWidth = 170,
  plotHeight = 150)
}
\arguments{
\item{GSrank}{piano GS rank object.}

\item{GS}{if GS = \code{"KEGG"}, the function will remove the "KEGG_" string in the GS name variable. Default is \code{"OTHER"}.}

\item{fileName}{Output file name.}

\item{plotWidth}{Set the width of the plot. Default is \code{170}.}

\item{plotHeight}{Set the height of the plot. Default is \code{150}.}
}
\value{
Outputs a \code{pdf} boxplot figure file with allGSA results.
}
\description{
Generate boxplot from piano GS rank object,
}
\details{
GSrank takes the resulted object generated from the piano function consensusScores().
}
\examples{
\dontrun{
pcPos_rank_mxdn <- consensusScores(pc_Pos, class = "mixed", direction="down",n = 15, adjusted = TRUE,
                                   method = "median", plot = TRUE, showLegend = F,
                                   rowNames = "names")

rbioArray_allGSbox(pcPos_rank_mxdn, fileName = "pcPos_rank_mxdn", plotWidth = 260, plotHeight = 240)

}
}

