% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rbioarray_hcluster}
\alias{rbioarray_hcluster}
\title{rbioarray_hcluster}
\usage{
rbioarray_hcluster(plotName = "data", fltlist, genesymbolOnly = FALSE,
  anno = NULL, genesymbolVar = NULL, n = "all", rmControl = TRUE,
  sampleName = NULL, fct, colGroup = ifelse(length(levels(fct)) < 19,
  length(levels(fct)), 19), distance = "euclidean", clust = "complete",
  colColour = "Paired", mapColour = "PRGn", n_mapColour = 11, ...,
  plotWidth = 7, plotHeight = 7)
}
\arguments{
\item{plotName}{File name for the export \code{pdf} plot file. Default is \code{"data"}.}

\item{fltlist}{Input filtered data, either a list, \code{EList} or \code{MAList} object.}

\item{anno}{Annotation object, usually a \code{dataframe}. Make sure to name the probe ID variable \code{ProbeName}. Only set this argument when \code{genesymbolVar = TRUE}. Default is \code{NULL}.}

\item{genesymbolVar}{Whether or not to remove probes without gene symbol. Default is \code{FALSE}.}

\item{n}{Number of genes to be clustered, numeric input or \code{"all"}. Default is \code{"all"}.}

\item{rmControl}{If to remove control probes (Agilent platform). Default is \code{TRUE}.#'}

\item{sampleName}{A \code{vector} containing names for column. Default is \code{NULL} and the function will use the column name from the input.}

\item{fct}{Input \code{factor} object for samples.}

\item{colGroup}{Colour group, numeric or dependent on \code{fct}.}

\item{distance}{Distance calculation method. Default is \code{"euclidean"}. See \code{\link{dist}} for more.}

\item{clust}{Clustering method. Default is \code{"complete"}. See \code{\link{hclust}} for more.}

\item{colColour}{Column group colour. Default is \code{"Paired"}. See \code{RColorBrewer} package for more.}

\item{mapColour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n_mapColour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{plotWidth}{Width of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{plotHeight}{Height of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{anno} object. Only set this argument when \code{genesymbolVar = TRUE}. Default is \code{NULL}.}
}
\value{
A heatmap based on hierarchical clustering analysis in \code{pdf} format.
}
\description{
Wrapper for hierarchical clustering analysis and heatmap visualization.
}
\examples{
\dontrun{

# standard usage
rbioarray_hcluster(fltlist = normdata, n = 500, fct = conSum, trace = "none", srtCol = 45,
offsetCol = 0, adjCol = c(1, 0), labRow = FALSE, key.title = "", keysize = 1.5,
key.xlab = "Normalized expression value", key.ylab = "Probe count")

# for non microarray or RNAseq data sets
###### unsupervised heatmap ######
## load the file
raw <- read.csv(file = "all_data.csv", na.strings = " ", stringsAsFactors = FALSE, check.names = FALSE)

## build the index
idx <- raw[, 1:2] # extract the sample information

conSum <- factor(idx$Condition, levels = unique(idx$Condition)) # extract the factor

# create the input data for RBioArray pacakge
rawT <- t(raw[, -(1:2)])
colnames(rawT) <- idx$SampleID
rawT <- apply(rawT, c(1,2), FUN = log2) # log2 tranforamtion
inputlist <- list(E = as.matrix(rawT),
                  genes = data.frame(GeneNames = rownames(rawT), ControlType = rep(0, length(rownames(rawT))),
                  stringsAsFactors = FALSE), target = idx)

rbioarray_hcluster(plotName = "all", fltlist = inputlist, n = "all", rmControl = FALSE, fct = conSum,
                   trace = "none",
                   distance = "euclidean", clust = "complete", colColour = "Paired",
                   mapColour = "RdBu", n_mapColour = 11,
                   srtCol = 30, offsetCol = 0, labRow = inputlist$genes$GeneNames,
                   key.title = "", keysize = 1.5, cexCol = 0.9, cexRow = 0.6,
                   key.xlab = "Log expression value", key.ylab = "miRNA count",
                   plotWidth = 10, plotHeight = 10)

}
}
