% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_PreProc}
\alias{rbioarray_PreProc}
\title{rbioarray_PreProc}
\usage{
rbioarray_PreProc(RawData, bgMethod = "auto", normMethod = "quantile", ...)
}
\arguments{
\item{RawData}{Input data, either a list, \code{EList} or \code{MAList} object.}

\item{...}{arguments for \code{backgroundCorrect.matrix()} or \code{backgroundCorrect()} functions from \code{limma} package.}

\item{BgMtd}{Background correction method. Default is \code{"auto"}. See \code{backgroundCorrect()} function from \code{limma} package for details.}

\item{BgOffst}{The offset to add to the background corrected values, as a base line. Default is \code{"50"}.}

\item{NormMtd}{Normalization method. Default is \code{"quantile"}. See \code{normalizeBetweenArrays()} function from \code{limma} package for details.}
}
\value{
Depending on the input type, the function outputs a \code{list}, \code{Elist} or \code{MAList} object with corrected and normalized expression values.
}
\description{
Data pre-processing function for the
}
\examples{
\dontrun{
normdata <- rbioarray_PreProc(mydata)
}
}

