% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_rlist.default}
\alias{rbioarray_rlist.default}
\title{rbioarray_rlist.default}
\usage{
\method{rbioarray_rlist}{default}(raw.dataframe,
  raw.background.signal.matrix = NULL, raw.annot.var.name = NULL,
  raw.gene_id.var.name = NULL, gene.annot.dataframe = NULL,
  gene.annot.gene_id.var.name = NULL,
  gene.annot.gene_symbol.var.name = NULL,
  gene.annot.control_type.var.name = NULL,
  gene.annot.control_type.val.pos = 1,
  gene.annot.control_type.val.neg = -1,
  gene.annot.control_type.val.exp = 0, gene.annot.rm.var.name = NULL,
  target.annot.file = NULL, target.annot.file.path = getwd(),
  sample_groups.var.name = NULL, verbose = TRUE)
}
\arguments{
\item{raw.dataframe}{Input data frame containing microarray hybridization signals with rows as probe/gene/genomic features and columns as samples. Note: the data frame should contain at least one annotation column.}

\item{raw.background.signal.matrix}{A opttional matrix containing background signals. The dimesnion should be the same as the input expression data without annotation columns.}

\item{raw.annot.var.name}{A string vector containing variable (i.e. column) name(s) for all the annotation columns in \code{raw.dataframe}.}

\item{raw.gene_id.var.name}{Variable (i.e. column) name for gene/probe/genomic feature identification from \code{raw.dataframe}.}

\item{gene.annot.dataframe}{Optional annotation data frame for gene/probe/genomic feature annotation.}

\item{gene.annot.gene_id.var.name}{Set only when \code{gene.annot.dataframe} is provided, variable name for probe/gene/genomic features identification from \code{gene.annot.dataframe}.}

\item{gene.annot.gene_symbol.var.name}{Set only when \code{gene.annot.dataframe} is provided, variable name for probe/gene/genomic features display name from \code{gene.annot.dataframe}, e.g. gene symbols.}

\item{gene.annot.control_type.var.name}{Optional and set only when \code{gene.annot.dataframe} is provided, name for the variable that denotes if the gene is a control and its control type. Default is \code{NULL}.}

\item{gene.annot.control_type.val.pos}{Set only when \code{gene.annot.control_type.var.name} is provided, value indicating a positive control probe. Default is \code{1}.}

\item{gene.annot.control_type.val.neg}{Set only when \code{gene.annot.control_type.var.name} is provided, value indicating a negative control probe. Default is \code{-1}.}

\item{gene.annot.control_type.val.exp}{Set only when \code{gene.annot.control_type.var.name} is provided, value indicating a none control probe. Default is \code{0}.}

\item{gene.annot.rm.var.name}{Optional variable names for the columns to remove from the gene annotation \code{genes} in the output.}

\item{target.annot.file}{File name for the target (i.e. sample) annotation \code{.csv} file.}

\item{target.annot.file.path}{The directory for \code{target.annot.file}. Default is \code{getwd()}.}

\item{sample_groups.var.name}{The variable name for sample groupping information from \code{target.annot.file}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\description{
The default \code{rbioarrary_rlist} function.
}
\details{
The \code{raw.background.signal.matrix} is usefual when processing a \code{EListRaw} class object from \code{limma} package.

         The word "gene" used in argument names and output item names is in its broader meaning of gene/probe/genomic feature.

         The optional annotation data frame \code{gene.annot.dataframe} should contain any additional annotation information in addition to the annotation column(s) from \code{raw.dataframe}.
         It is noted that \code{gene.annot.dataframe} should contain at least one column for the sample type of gene/probe/genomic feature identification as the identification variable from  \code{raw.dataframe}.
         Such column is set via argument \code{raw_file.gene_id.var_name} and \code{genes_annotation.gene_id.var_name}.
         The gene display name will only be used if \code{gene.annot.dataframe} is used. Otherwise, it wil use \code{raw.gene_id.var.name} from \code{raw.dataframe}.
}
