% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_rlist.EListRaw}
\alias{rbioarray_rlist.EListRaw}
\title{rbioarray_rlist.EListRaw}
\usage{
\method{rbioarray_rlist}{EListRaw}(object, ...)
}
\arguments{
\item{object}{A input \code{EListRaw} class object from \code{limma} package}

\item{...}{Additional argument for the default function.}
}
\description{
The \code{rbioarrary_rlist} function for \code{EListRaw} class
}
