% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_transfo_normalize.default}
\alias{rbioarray_transfo_normalize.default}
\title{rbioarray_transfo_normalize.default}
\usage{
\method{rbioarray_transfo_normalize}{default}(E, E.background = NULL,
  bgc.method = "auto", between.sample.norm.method = "quantile",
  between.sample.weight.design = NULL, ..., verbose = TRUE)
}
\arguments{
\item{E}{Input raw expression value matrix with columns for samples, rows for genes/probes/genomic features.}

\item{E.background}{A opttional matrix containing background signals. The dimesnion should be the same as the input expression data without annotation columns.}

\item{bgc.method}{Background correction method. Default is \code{"auto"}. See \code{backgroundCorrect()} function from \code{limma} package for details.}

\item{between.sample.norm.method}{Normalization method. Default is \code{"quantile"}. See \code{normalizeBetweenArrays()} function from \code{limma} package for details.}

\item{between.sample.weight.design}{Microarray experiment sample design matrix for beteween sample weight calculation.}

\item{...}{Additional arguments the default method \code{backgroundCorrect.matrix} function from \code{limma} package.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
A list with the core items for a \code{rbioarray_plist} class.
}
\description{
\code{\link{rbioarray_transfo_normalize}} for \code{rbioarray_rlist} class object.
}
\details{
The \code{rbioarray_rlist} object can be obtained from \code{\link{rbioarray_rlist}} function.
         The word "gene" used in argument names and output item names is in its broader meaning of gene/probe/genomic feature.
         ADD the fact that \code{normalizeBetweenArrays} function is where \code{limma} log transforms data.
}
