% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_transfo_normalize.rbioarray_rlist}
\alias{rbioarray_transfo_normalize.rbioarray_rlist}
\title{Title rbioarray_transfo_normalize.rbioarray_rlist}
\usage{
\method{rbioarray_transfo_normalize}{rbioarray_rlist}(object, design, ...,
  verbose = TRUE)
}
\arguments{
\item{object}{Input obejct with raw data and annotation information. Could be \code{rbioarray_rlist}, \code{Elist} or \code{MAList} classes.}

\item{design}{Microarray experiment sample design matrix. Make sure the design colnames are the same as the levles of \code{object$sample_groups}.}

\item{...}{Additional arguments the default method \code{\link{rbioarray_transfo_normalize.default}}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
A \code{rbioarray_plist} class object.
}
\description{
\code{\link{rbioarray_transfo_normalize}} for \code{rbioarray_rlist} class object.
}
\details{
The \code{rbioarray_rlist} object can be obtained from \code{\link{rbioarray_rlist}} function.
}
