% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS}
\alias{rbioGS}
\title{rbioGS}
\usage{
rbioGS(
  GS = NULL,
  GSfile = NULL,
  idVar = NULL,
  method_p = c("fisher", "stouffer", "reporter", "tailStrength", "wilcoxon"),
  pVar = NULL,
  logFCVar = NULL,
  method_t = c("page", "gsea", "maxmean"),
  tVar = NULL,
  ...,
  parallelComputing = FALSE,
  clusterType = "PSOCK"
)
}
\arguments{
\item{GS}{pre-loaded gene set objects. Default is \code{NULL}.}

\item{GSfile}{GS database file. Set only if \code{GS} argument is \code{NULL}. File format should be \code{gmt}. If the working directory isn't set, make sure to include the full path. Default is \code{NULL}.}

\item{idVar}{Gene IDs. Could be, but not exclusive to, a variable of a dataframe. Must be the same length as \code{pVar}, \code{logFCVar} and \code{tVar}. Currently only takes \code{Entrez ID}.}

\item{method_p}{Gene set enrichment methods that takes \code{p value} and \code{logFC}. Default is \code{c("fisher", "stouffer", "reporter", "tailStrength", "wilcoxon")}, and can be set as \code{NULL}.}

\item{pVar}{Set only method_p is not NULL, DE p values. Could be a variable of a dataframe, or a vector. Must be the same length as \code{logFCVar}, \code{tVar} and \code{idVar}. Can be set as \code{NULL}.}

\item{logFCVar}{Set only method_p is not NULL, DE logFC (log fold change). Could be a variable of a dataframe, or a vector. Must be the same length as \code{pVar}, \code{tVar} and \code{idVar}. Can be set as \code{NULL}.}

\item{method_t}{Gene set enrichment methods that takes \code{t statistics}. Default is \code{c("page", "gsea", "maxmean")}, and can be set as \code{NULL}.}

\item{tVar}{Set only method_t is not NULL, DE t values. Gene level t values. Could be a variable of a dataframe, or a vector. Must be the same length as \code{pVar}, \code{logFCVar} and \code{idVar}. Can be set as \code{NULL}.}

\item{...}{Arguments to pass to \code{runGSA} function from \code{piano} package. See the corresponding help page from of \code{piano} for details.}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs a \code{list} object with GSA results, both for p value based and t value based.
}
\description{
gene set analysis using piano
}
\details{
The function is based on piano package. It runs "fisher", "stouffer", "reporter", "tailStrength", "wilcoxon" for p value based GSA, and "page", "gsea", "maxmean" for t value based GSA.
}
\examples{
\dontrun{
gsoutput <- rbioGS(GS = kegg, pVar = dfm$p_value, logFCVar = dfm$logFC, tVar = dfm$t_value, idVar = dfm$EntrezID, nPerm = 1000, parallelComputing = TRUE, clusterType = "FORK")


}
}
