% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_all_noplot}
\alias{rbioGS_all_noplot}
\title{rbioGS_all_noplot}
\usage{
rbioGS_all_noplot(input, entrezVar = NULL, GS = NULL, ...,
  parallelComputing = FALSE, cluster = "PSOCK")
}
\arguments{
\item{input}{The input list with DE reuslt, i.e. a collection of dataframes from \code{topTable} function of \code{limma} package.}

\item{entrezVar}{Name of the EntrezID variable in the \code{input} object.}

\item{GS}{Pre-loaded gene set objects.}

\item{...}{Arguments to pass to \code{\link{rbioGS}}.}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}}

\item{cluster}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs  \code{csv} files and \code{pdf} figure files with GSA results.
}
\description{
All-in-one wrapper for GSA (no plotting).
}
\details{
This is an all-in-one function for GS anlyasis based on piano package. It runs "fisher", "stouffer", "reporter", "tailStrength", "wilcoxon" for p value based GSA, and "page", "gsea", "maxmean" for t value based GSA (customizable). See arguments for \code{\link{rbioGS}} for details.
}
\examples{
\dontrun{

rbioGS_all_noplot(input = DElist, entrezVar = "EntrezID", method_p = c("stouffer", "fisher"), method_t = c("gsea", "maxmean"), nPerm = 1000, GS = kegg, parallelComputing = TRUE, cluster = "FORK")

}
}

