% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_kegg}
\alias{rbioGS_kegg}
\title{rbioGS_kegg}
\usage{
rbioGS_kegg(dfm, entrezVar = NULL, keggID, suffix, species = "hsa", ...)
}
\arguments{
\item{dfm}{GS dataframe with \code{ENTREZID} and \code{logFC} variables.}

\item{entrezVar}{Name of the EntrezID variable in the \code{DElst} object.}

\item{keggID}{Make sure to have quotation marks around the ID number.}

\item{suffix}{Output file name suffix. Make sure to put it in quotation marks.}

\item{species}{Set the species. Default is \code{"hsa"}. Visit kegg website for details.}

\item{...}{Additinal arguments for \code{pathview()} from \code{pathview} package.}
}
\value{
Outputs a \code{list} kegg object, as well as masked figure files in \code{pdf} format.
}
\description{
Download and generate DE results masked kegg pathway figures, using pathview package
}
\details{
GSrank takes the resulted object generated from the piano function consensusScores().

Visit website \url{http://www.genome.jp/kegg/pathway.html} for kegg IDs.
}
\examples{
\dontrun{

 rbioGS_kegg(all_DE4GSA_flt, keggID = "04060") # Cytokine-cytokine receptor interaction, disdn + mixdn + nd

}
}
