% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_scatter}
\alias{rbioGS_scatter}
\title{rbioGS_scatter}
\usage{
rbioGS_scatter(
  GSA_list,
  fileName = "GS_list",
  ...,
  plotTitle = NULL,
  xLabel = "median p value",
  yLabel = "log consensus score",
  rankCutoff = 20,
  pCutoff = 0.05,
  plotWidth = 170,
  plotHeight = 150
)
}
\arguments{
\item{GSA_list}{piano GS results object.}

\item{fileName}{Output file name.}

\item{...}{Arguments passing to \code{consensusHeatmap} function from \code{piano} package. See the responding help page of \code{piano} for details.}

\item{plotTitle}{Title of the plot. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Default is \code{"median p value"}.}

\item{yLabel}{Y-axis label. Default is \code{"log consensus score"}.}

\item{rankCutoff}{Cutoff value for GS rank line. The input number will be log2 transformed when plotting. Default is \code{20}.}

\item{pCutoff}{Cutoff value for GS p value line. Default is \code{0.05}.}

\item{plotHeight}{Set the height of the plot. Default is \code{150}.}
}
\value{
Outputs a \code{pdf} scatter figure file with allGSA results.
}
\description{
Generate scatter plot for piano GS rank heatmap obejct. Note that the rank is log2 transformed.
}
\details{
The function is based on piano package.
}
\examples{
\dontrun{

rbioGS_scatter(GSA_list = GS_Pos, fileName = "GS_pos", cutoff = 15, method = "median", adjusted = TRUE, rankCutoff = 50, pCutoff = 0.05)

}
}
