% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_sp2hsaEntrez}
\alias{rbioGS_sp2hsaEntrez}
\title{rbioGS_sp2hsaEntrez}
\usage{
rbioGS_sp2hsaEntrez(
  DElst,
  tgtSpecies = c("mmu", "mouse", "rno", "rat", "medaka", "olatipes"),
  ensemblTransVar = NULL,
  entrezVar = NULL,
  parallelComputing = FALSE,
  clusterType = "PSOCK"
)
}
\arguments{
\item{DElst}{The list with DE reuslt, from functions \code{\link{rbioarray_DE}} or \code{\link{rbioseq_DE}}.}

\item{tgtSpecies}{The target species. Options are \code{"mmu"}, \code{"rno"} and \code{"medaka"}.}

\item{ensemblTransVar}{The name of the variable from DE list containing ensembl transcript ID.}

\item{entrezVar}{The name of the variable from DE list containing entrez gene ID.}

\item{parallelComputing}{If to use parallel computing. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs a DE \code{list} object with human Entrez ID for each dataframe. This list has the exact same format as the input DE list.
}
\description{
Using up-to-date ensembl database, convert from mouse/rat ensemble transcript ID to and add Human entrez ID to the DE list from the DE functions, i.e. \code{\link{rbioarray_DE}} or \code{\link{rbioseq_DE}}, as human EntrezID is needed for GS analysis if using human gene sets.
}
\details{
IMPORTANT: this function requires an internet connection as it retrieves information from ensembl website for human gene orthologs.
}
\examples{
\dontrun{
rbioGS_sp2hsaEntrez(DElst = comparison_DE, tgtSpecies = "mmu", ensemblTransVar = "EnsemblID")
}
}
