% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_table}
\alias{rbioGS_table}
\title{rbioGS_table}
\usage{
rbioGS_table(GSA_list, fileName = "GS_list", export.name = NULL, ...)
}
\arguments{
\item{GSA_list}{piano GS results object.}

\item{fileName}{Output file name.}

\item{export.name}{String for export file and object prefix. Default is \code{NULL}.}

\item{...}{Arguments passing to \code{consensusHeatmap} function from \code{piano} package. See the responding help page of \code{piano} for details.}
}
\value{
Outputs a \code{pdf} scatter figure file, an R data frame as well as a \code{csv} summary with allGSA results.
}
\description{
Generate table for piano GS resutls. Note that the rank is log2 transformed.
}
\details{
The function is based on piano package.
}
\examples{
\dontrun{

rbioGS_table(GSA_list = GS_Pos, fileName = "GS_pos", cutoff = 15, method = "median", adjusted = TRUE, rankCutoff = 50, pCutoff = 0.05)

}
}
