% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDEA.R
\name{rbioHtmap_unsuperv}
\alias{rbioHtmap_unsuperv}
\title{rbioseq_Htmap_unsuperv}
\usage{
rbioHtmap_unsuperv(rawDFM, voomNrm = TRUE, RNAtype = "mRNA",
  nPlot = "all", cexRow = 0.6, cexCol = 0.7)
}
\arguments{
\item{voomNrm}{If the data needs to be voom normalized. Default is \code{TRUE}.}

\item{RNAtype}{RNAtype takes \code{"mRNA"}, \code{"miRNA"} and \code{"lncRNA"}. Default is \code{"mRNA"}.}

\item{nPlot}{sets the number of the features to be plotted in the heatmap, and takes \code{"all"} or a numeric value. Default is \code{"all"}.}

\item{cexRow}{Font size. Default is \code{0.6}.}

\item{cexCol}{Also font size. Default is \code{0.7}.}

\item{AnnCountDfm}{Input data frame with annoatation merged.}
}
\value{
Outputs a heatmap based on the input dataframe.
}
\description{
unsupvised heatmap: with or without voom normalization
}
\examples{
\dontrun{
DE_dataframe <- rbioArray_DE(dataframe)
}
}

