% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_count_to_de.R
\name{rbioseq_clr_ilr_transfo}
\alias{rbioseq_clr_ilr_transfo}
\title{rbioseq_clr_ilr_transfo}
\usage{
rbioseq_clr_ilr_transfo(
  x,
  offset = 1,
  mode = c("clr", "ilr"),
  ilr.method.fast = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{Input read count data matrix.}

\item{offset}{Read count offset value added to avoid zero. Default is \code{1}.}

\item{mode}{Log ratio transformation method. Options are "clr" (centered log transformation) and "ilr" (isometric log transformation). Default is \code{"clr"}.}

\item{ilr.method.fast}{Useful only when \code{mode = "ilr"}. Default is \code{TRUE}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
A data matrix with log ratio transformed values.
}
\description{
Log ratio transformation function for read count data. Row: sample, column: features.
}
\details{
This function is needed as part of the data pre-processing procedure to run multivariate and machine learning analysis featured in \code{RBioFS} package.

         As per Quinn et al. (2018), NGS data can be considered as compositional data. As such, data must be transformed for usual statistical analysis and visualization.
         Log ratio transformation serves such purpose. Note that the number of features will be one less when using "ilr" method.

         It is not to be combined with the other normalization methods featured in the \code{\link{rbioseq_DE}}.

         Ref: Quinn TP, et al. 2018. Understanding sequencing data as compositions: an outlook and review. Bioinformatics. 2018: 1 - 9.
}
\examples{
\dontrun{
tstX <- rbioseq_clr_ilr_transfo(tstdata, offset = 1, mode = "clr")
}
}
