% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDEA.R
\name{rbioseq_gtf}
\alias{rbioseq_gtf}
\title{Title rbioseq_gtf}
\usage{
rbioseq_gtf(
  file,
  verbose = TRUE,
  parallelComputing = FALSE,
  clusterType = "FORK",
  n_cores = NULL
)
}
\arguments{
\item{file}{GTF/GFF file. Add path if needed.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{n_cores}{Only set when \code{parallelComputing = TRUE}, number of cores to use. When \code{NULL}, the function uses max-1.}

\item{cluterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
A data frame with items from GTF/GFF file.
}
\description{
Import GTF/GFF files
}
\details{
The following items are extracted from GTF/GFF file: \code{gene_id}, \code{gene_name}, \code{transcript_id}, \code{transcript_name}, \code{chr}, \code{type}, .
}
\examples{

\dontrun{
gtf <- rbioseq_gtf(file = "gencode_v19.gtf")
}

}
