% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacyFunctions.R
\name{rbioseq_hcluster}
\alias{rbioseq_hcluster}
\title{rbioseq_hcluster}
\usage{
rbioseq_hcluster(plotName = "data", dfm_count = NULL,
  dfm_annot = NULL, geneidVar = "gene_id", count_threshold = "none",
  design = NULL, qc_plot = FALSE, n = "all", sampleName = NULL,
  fct = NULL, colGroup = ifelse(length(levels(fct)) < 19,
  length(levels(fct)), 19), distance = c("euclidean", "maximum",
  "manhattan", "canberra", "binary", "minkowski"), clust = c("complete",
  "ward.D", "ward.D2", "single", "average", "mcquitty", "median",
  "centroid"), colColour = "Paired", mapColour = "PRGn",
  n_mapColour = 11, ..., plotWidth = 7, plotHeight = 7)
}
\arguments{
\item{plotName}{File name for the export \code{pdf} plot file. Default is \code{"data"}.}

\item{dfm_count}{Dataframe contains the feature read counts, with rows as genomic featues (or genes) and column as samples. Default is \code{NULL}.}

\item{dfm_annot}{Dataframe contains the gene annotation information, with rows as genmic features and columns as annotation variables. The row lengths of this dataframe should be the same as \code{dfm_count}.}

\item{count_threshold}{Read count threshold. No filtering will be applied when set \code{"none"}. Otherwise, a numeric number can be set as the minimum read count for filtering. DDefault is \code{"none"}.}

\item{qc_plot}{Wether or not to produce a QC plot upon filtering, normalization and weight calculation. Default is \code{FALSE}.}

\item{n}{Number of genes to be clustered, numeric input or \code{"all"}. Default is \code{"all"}.}

\item{sampleName}{A \code{vector} containing names for column. Default is \code{NULL} and the function will use the column name from the input.}

\item{fct}{Input \code{factor} object for samples. Default is \code{NULL}.}

\item{colGroup}{Colour group, numeric or dependent on \code{fct}.}

\item{distance}{Distance calculation method. Default is \code{"euclidean"}. See \code{\link{dist}} for more.}

\item{clust}{Clustering method. Default is \code{"complete"}. See \code{\link{hclust}} for more.}

\item{colColour}{Column group colour. Default is \code{"Paired"}. See \code{RColorBrewer} package for more.}

\item{mapColour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n_mapColour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{plotWidth}{Width of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{plotHeight}{Height of the plot. Unit is \code{inch}. Default is \code{7}.}
}
\value{
A heatmap based on hierarchical clustering analysis in \code{pdf} format
}
\description{
Legacy function. Wrapper for hierarchical clustering analysis and heatmap visualization for RNA seq data.
}
\details{
The data filtering and normalization functions are also included in the function.
}
\examples{
\dontrun{

# standard usage

}
}
