% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDEA.R
\name{rbioseq_import_gtf}
\alias{rbioseq_import_gtf}
\title{Title rbioseq_import_gtf}
\usage{
rbioseq_import_gtf(file, verbose = TRUE)
}
\arguments{
\item{file}{GTF/GFF file. Add path if needed.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{parallelComputing}{Wether to use parallel computing or not. Default is \code{TRUE}.}

\item{cluterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
A data frame with items from GTF/GFF file.
}
\description{
Import GTF/GFF files
}
\details{
The following items are extracted from GTF/GFF file: \code{chromosome}, \code{gene_id}, \code{gene_type}, and \code{gene_name}.
}
\examples{

\dontrun{
gtf <- rbioseq_import_gtf(file = "gencode_v19.gtf")
}

}
