% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_count_to_de.R
\name{rnaseq_de.default}
\alias{rnaseq_de.default}
\title{Title rnaseq_de.default}
\usage{
\method{rnaseq_de}{default}(x, y = NULL, y.gene_id.var.name = "genes",
  y.gene_symbol.var.name = "genes",
  filter.threshold.cpm.count = "none",
  filter.threshold.min.sample = NULL, annot.group = NULL,
  between.genes.norm.method = "TMM", design, contra, qc.plot = TRUE)
}
\arguments{
\item{x}{Input read count matrix, with rows for genes and columns for RNA samples}

\item{y}{Target (e.g. genes) annotation matrix or vector.}

\item{y.gene_id.var.name}{Variable name for gene (i.e. target) identification. Default is \code{"genes"}.}

\item{y.gene_symbol.var.name}{Variable name for gene (i.e. target) symbols. Default is \code{"genes"}.}

\item{filter.threshold.cpm.count}{Filtering threshold for counts based on CPM (counts per million). Default is \code{"none"}.}

\item{filter.threshold.min.sample}{Minimum number of samples meeting the count threshold. Default is \code{NULL}.}

\item{annot.group}{Set only when \code{filter.threshold.min.sample} is not \code{NULL}, Sample group annotation object. Can be a \code{factor} or \code{vector} object.}

\item{between.genes.norm.method}{Between gene normalization method. Options are: \code{"none"}, \code{"TMM"}, \code{"RLE"}, \code{"upperquartile"}. Default is \code{"TMM"}.}

\item{design}{Design matrix.}

\item{contra}{Contrast matrix.}

\item{qc.plot}{QC plot for the input read counts}
}
\value{
A \code{rbioseq_de} object. The items of the object are following:

\code{filter_results}

\code{normalization_results}

\code{gene_id_var_name}

\code{gene_symbol_var_name}

\code{F_stats}

\code{DE_results}
}
\description{
The default \code{rnaseq_de} function
}
