% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_count_to_de.R
\name{rnaseq_de.rbioseq_count}
\alias{rnaseq_de.rbioseq_count}
\title{Title rnaseq_de.rbioseq_count}
\usage{
\method{rnaseq_de}{rbioseq_count}(object,
  filter.threshold.min.count = 10, ...)
}
\arguments{
\item{object}{A \code{rbioseq_count} object from \code{\code{rbioseq_import_count}} function.}

\item{filter.threshold.min.count}{Minimum count for the smallest library for filter thresholding. Default is \code{10}.}

\item{...}{Additional arguments for \code{\link{rnaseq_de.defuault}}.}
}
\description{
The \code{rnaseq_de} function for \code{rbioseq_count} object from \code{\code{rbioseq_import_count}} function.
}
