% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test.R
\name{sig.default}
\alias{sig.default}
\title{Title sig.default}
\usage{
\method{sig}{default}(input.de.list, input.gene_symbol.var.name,
  input.Fstats.matrix, experiment = NULL, FC = 1.5, alpha = 0.05,
  FDR = TRUE, gene_symbol = TRUE, export.name = NULL,
  export.mode = "all", plot = TRUE, plot.top.gene = FALSE,
  plot.top.gene.n = 5, plot.top.gene.padding = 0.5,
  plot.Title = NULL, plot.xLabel = "log2(fold change)",
  plot.yLabel = "-log10(p value)", plot.symbolSize = 2,
  plot.sigColour = "red", plot.nonsigColour = "gray",
  plot.xTxtSize = 10, plot.yTxtSize = 10, plot.Width = 170,
  plot.Height = 150)
}
\arguments{
\item{input.de.list}{Input list cantaining DE dataframes for each comparison.}

\item{input.gene_symbol.var.name}{Input gene sysmbol variable name from teh DE dataframes.}

\item{input.Fstats.matrix}{Input dataframe containing F stats}

\item{experiment}{Character string describing the experiment used to generate data, e.g. "microarray", "RNAseq".}

\item{FC}{Threshold for fold change. Default is \code{1.5}.}

\item{alpha}{Threshold for p values. Default is \code{0.05}.}

\item{FDR}{If to use FDR correction for p values. Default is \code{TRUE}.}

\item{gene_symbol}{If to apply gene symbols in the plot and exported results. Default is \code{TRUE}.}

\item{export.name}{Name used for output objects to the environment and directory. Not optional. Default is \code{NULL}.}

\item{export.mode}{Mode used to export results to the directory. Options are \code{"all"}, \code{"all.gene_symbol"}, \code{"sig"}. Default is \code{"all"}. See details.}

\item{plot}{If to plot volcano plot. Default is \code{TRUE}.}

\item{plot.top.gene}{If to display the top gene identification, i.e., probem name or gene name, on the plot. Default is \code{FALSE}.}

\item{plot.top.gene.n}{When \code{plot.top.gene = TRUE}, to set how many genes to display. Default is \code{5}.}

\item{plot.top.gene.padding}{When \code{plot.top.gene = TRUE}, to set the distance between the dot and the gene symbol. Default is \code{0.5}.}

\item{plot.Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{plot.xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{plot.yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{plot.symbolSize}{Size of the symbol. Default is \code{2}.}

\item{plot.sigColour}{Colour of the significant genes or probes. Default is \code{"red"}.}

\item{plot.nonsigColour}{Colour of the non-significant genes or probes. Default is \code{"gray"}.}

\item{plot.xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{plot.yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plot.Width}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plot.Height}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{annot} object. Only set this argument when \code{geneName = TRUE}. Default is \code{NULL}.}
}
\value{
Signifiance test results as \code{csv} files to the working directory, as well as a \code{sig} object to the environment.
        The \code{sig} object contains the following items;

        \code{significant_change_summary}

        \code{export.mode}

        \code{signifianct_change_results}: this only contains significantly changed genes.

        \code{experiment}
}
\description{
The default \code{sig} function.
}
\details{
Explanation for \code{export.mode} options:

        \code{all}: export all results for all probes/features with or without name (e.g. gene symbol, gene name, etc.) annotations.

        \code{all.gene_symbol}: export all results for probes/features only with name (e.g. gene symbol, gene name, etc.) annotations.

        \code{sig}: export only the signifiant changes. Gene symbol settings depends on argument \code{gene_symbol}.
}
