% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test.R
\name{sig.rbioarray_de}
\alias{sig.rbioarray_de}
\title{sig.rbioarray_de}
\usage{
\method{sig}{rbioarray_de}(object, p.val.correction.method = c("fdr",
  "spikein", "none"), export.name = NULL, ...)
}
\arguments{
\item{object}{A \code{rbioarray_de} object \code{\link{microarray_de}} function.}

\item{p.val.correction.method}{A character string describing the p value correction method used for significant test. Options are \code{"fdr"}, \code{"spikein"} and \code{"none"}. Default is \code{"fdr"}.}

\item{export.name}{Optional name used for output objects to the environment and directory. Default is \code{NULL}.}

\item{...}{Additional arguments for \code{\link{sig.defuault}}.}
}
\description{
The \code{sig} function for \code{rbioarray_de} object from \code{\link{microarray_de}} function.
}
