% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{microarray_de}
\alias{microarray_de}
\title{microarray_de}
\usage{
microarray_de(object, contra, verbose = TRUE)
}
\arguments{
\item{object}{The input \code{rbioarray_flist} object.}

\item{contra}{contra Contrast matrix.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
A \code{rbioarray_de} class object with DE results, containing the following core items:

        \code{F_stats}

        \code{DE_results}: a list containing the DE results

        \code{comparisons}: a list with comparisons and comparison levels

        \code{fit}: the limma fitted DE object as a reference

        \code{input_data}: input E matrix and genes data frame from rbioarray_flist object
                           note the E matrix is \code{E} from the \code{rbioarray_flist} object, meaning the filtered and normalized data

        Additionally, the \code{rbioarray_de} includes additional items from the input \code{rbioarray_flist} object.
}
\description{
Function that performs statistical analysis for microarray data from \code{rbioarray_flist} class object.
}
\examples{
\dontrun{

de_list <- microarray_de(object = filtered_list, contra = contra)

}
}
