% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rbio_network}
\alias{rbio_network}
\alias{rbio_network.rbio_tom_graph}
\alias{rbio_network.default}
\title{rbio_network}
\usage{
rbio_network(object, ...)

\method{rbio_network}{rbio_tom_graph}(object, export.name = NULL, ...)

\method{rbio_network}{default}(
  g,
  membership = NULL,
  export.name = NULL,
  colour_scheme = c("Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2",
    "Set3"),
  initial_colour_number = 8,
  plot.title = "Network",
  plot.margins = c(5, 5, 5, 5),
  plot.highlight_membership = TRUE,
  plot.layout_type = "circular",
  plot.vertex.size = NA,
  plot.vertex.sizeScale = c(1, 4),
  plot.vertex.label.size,
  plot.vertex.label.color = "black",
  plot.vertex.label.dist = 0,
  plot.edge.filter = 0.95,
  plot.edge.weightScale = c(1, 4),
  plot.edge.arrow.mode = FALSE,
  plot.edge.curved = FALSE,
  plot.ellipse = FALSE,
  plot.height,
  plot.width,
  random_state = 1,
  verbose = TRUE
)
}
\arguments{
\item{object}{An \code{rbio_tom_graph} object.}

\item{...}{Additional arguments for corresponding S3 class methods.}

\item{export.name}{string. Optional prefix for export file name. Default is \code{NULL}.}

\item{g}{Input list cantaining DE dataframes for each comparison.}

\item{membership}{Input dataframe containing F stats.}

\item{colour_scheme}{string. The colour set from \code{\link{RColorBrewer}} function. Default is \code{"Accent"}.}

\item{initial_colour_number}{int. The number of starting colours to use for the clusters. See details. Default is \code{8}.}

\item{plot.highlight_membership}{boolean. <TBC: under construction>}

\item{plot.layout_type}{string. <TBC: under construction>}

\item{plot.vertex.size}{numeric. <TBC: under construction>}

\item{plot.vertex.sizeScale}{numeric two-vector. <TBC: under construction>}

\item{plot.vertex.label.size}{numeric. <TBC: under construction>}

\item{plot.vertex.label.dist}{numeric. <TBC: under construction>}

\item{plot.edge.filter}{numeric: 0~1. <TBC: under construction>}

\item{plot.edge.weightScale}{numeric two-vector. <TBC: under construction>}

\item{plot.edge.arrow.mode}{boolean. <TBC: under construction>}

\item{plot.edge.curved}{boolean. <TBC: under construction>}

\item{plot.ellipse}{boolean. <TBC: under construction>}

\item{plot.height}{numeric. <TBC: under construction>}

\item{plot.width}{numeric. <TBC: under construction>}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{plot.title.}{string. <TBC: under construction>}

\item{plot.margins.}{numeric four-vector. <TBC: under construction>}

\item{plot.vertex.label.colour}{string vector. <TBC: under construction>}
}
\value{
TBC

TBC
}
\description{
(UNDER CONSTRUCTION: NOT FUNCTIONAL) Network construction and visualization function
}
\details{
<TBC: under construction>
         For \code{colour_scheme}, use the following as a guide (name, maximum number of colours):
           Accent	8
           Dark2	8
           Paired	12
           Pastel1	9
           Pastel2	8
           Set1	9
           Set2	8
          Set3	12
         NOTE: The maximum number of colours does not reflect the number of clusters - it is simply what \code{\link{RColorBrewer}} requires.
               The recommended approach is to set \code{n_colours} to this number.
}
\examples{

\dontrun{
TBC
}


\dontrun{
TBC
}

}
