% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rbio_network}
\alias{rbio_network}
\alias{rbio_network.rbio_tom_graph}
\alias{rbio_network.default}
\title{rbio_network}
\usage{
rbio_network(object, ...)

\method{rbio_network}{rbio_tom_graph}(object, export.name = NULL, ...)

\method{rbio_network}{default}(
  g,
  export.name = NULL,
  g_membership = NULL,
  colour_scheme = c("Accent", "Dark2", "Paired", "Pastel1", "Pastel2", "Set1", "Set2",
    "Set3"),
  initial_colour_number = 8,
  plot.title = "Network",
  plot.margins = c(5, 5, 5, 5),
  plot.font.family = "sans",
  plot.highlight_membership = TRUE,
  plot.layout_type = c("circular", "fr", "tree", "nicely", "sphere"),
  plot.vertex.remove.zerodegree = TRUE,
  plot.vertex.size = NULL,
  plot.vertex.size.scale = c(1, 4),
  plot.vertex.label = NULL,
  plot.vertex.label.topvsize = TRUE,
  plot.vertex.label.topvsize.percent = 0.05,
  plot.vertex.color.highlighttopvsize = TRUE,
  plot.vertex.label.size = 0.5,
  plot.vertex.label.color = "black",
  plot.vertex.label.dist = 0,
  plot.edge.filter = 0.05,
  plot.edge.weight = NULL,
  plot.edge.weight.scale = c(1, 4),
  plot.edge.arrow.mode = FALSE,
  plot.edge.curved = FALSE,
  plot.edge.color.highlighttopvsize = TRUE,
  plot.ellipse = FALSE,
  plot.height = 7,
  plot.width = 7,
  random_state = 1,
  verbose = TRUE
)
}
\arguments{
\item{object}{An \code{rbio_tom_graph} object.}

\item{...}{Additional arguments for corresponding S3 class methods.}

\item{export.name}{string. Optional prefix for export file name. Default is \code{NULL}.}

\item{g}{Input list cantaining DE dataframes for each comparison.}

\item{g_membership}{Input dataframe containing F stats.}

\item{colour_scheme}{string. The colour set from \code{\link{RColorBrewer}} function. Default is \code{"Accent"}.}

\item{initial_colour_number}{int. The number of starting colours to use for the clusters. See details. Default is \code{8}.}

\item{plot.font.family}{string. The font family of the labels in the plot. Default is \code{"sans"}.}

\item{plot.highlight_membership}{boolean. <TBC: under construction>}

\item{plot.layout_type}{string. <TBC: under construction>}

\item{plot.vertex.remove.zerodegree}{boolean. If to remove zero degree edges from the graph. Default is \code{TRUE}.}

\item{plot.vertex.size}{numeric vector. <TBC: under construction>}

\item{plot.vertex.size.scale}{numeric two-vector. <TBC: under construction>}

\item{plot.vertex.label}{string vector. Optional custom vertex label. Default is \code{NULL}, meaning V(g)$name.}

\item{plot.vertex.label.topvsize}{Boolean. If to display labels with a threshold on vertex size. \code{default is FALSE}.}

\item{plot.vertex.color.highlighttopvsize}{Boolean. When \code{plot.vertex.label.topvsize = TRUE}, if to make non-top vertices transparent and frameless. Default is \code{TRUE}.}

\item{plot.vertex.label.size}{numeric vector. <TBC: under construction>}

\item{plot.vertex.label.dist}{numeric. <TBC: under construction>}

\item{plot.edge.filter}{numeric: 0-1. Top percentage edges to keep. Default is \code{0.05}.}

\item{plot.edge.weight}{numeric vector. <TBC: under construction>}

\item{plot.edge.weight.scale}{numeric two-vector. <TBC: under construction>}

\item{plot.edge.arrow.mode}{Boolean. <TBC: under construction>}

\item{plot.edge.curved}{Boolean. <TBC: under construction>}

\item{plot.edge.color.highlighttopvsize}{Boolean. When \code{plot.vertex.label.topvsize = TRUE}, if to make all edges transparent. Default is \code{TRUE}.}

\item{plot.ellipse}{Boolean. <TBC: under construction>}

\item{plot.height}{numeric. <TBC: under construction>}

\item{plot.width}{numeric. <TBC: under construction>}

\item{random_state}{integer. <TBC: under construction>}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}

\item{plot.title.}{string. <TBC: under construction>}

\item{plot.margins.}{numeric four-vector. <TBC: under construction>}

\item{plot.vertex.label.topvsize.filter}{numeric: 0-1. Set when \code{plot.vertex.label.topvsize = TRUE}, top percetage size to display the vertex labels. Default is \code{0.05}.}

\item{plot.vertex.label.colour}{string vector. <TBC: under construction>}
}
\value{
TBC

TBC
}
\description{
(UNDER CONSTRUCTION: NOT FUNCTIONAL) Network construction and visualization function
}
\details{
<TBC: under construction>
         For \code{colour_scheme}, use the following as a guide (name, maximum number of colours):
           Accent	8
           Dark2	8
           Paired	12
           Pastel1	9
           Pastel2	8
           Set1	9
           Set2	8
           Set3	12
         NOTE: The maximum number of colours does not reflect the number of clusters - it is simply what \code{\link{RColorBrewer}} requires.
               The recommended approach is to set \code{initial_colour_number} to this number.

         The \code{plot.margins} follow the base R setting in \code{\link{par}} for the positioning:
         b: mar[1], l: mar[2], t: mar[3], r: mar[4]

         For \code{plot.vertex.label.topvsize.filter}, the functional will apply the filter per group if g_membership is provided.
}
\examples{

\dontrun{
TBC
}


\dontrun{
TBC
}

}
