% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_supervised_corcluster}
\alias{rbio_supervised_corcluster}
\title{rbio_supervised_corcluster}
\usage{
rbio_supervised_corcluster(object, gene_symbol.only = FALSE,
  cor.method = "pearson", map.colour = "PRGn", n.map.colour = 11,
  ..., heatmap.width = 7, heatmap.height = 7, sigplot.adj.p = FALSE,
  sigplot.alpha = 0.05, sigplot.sigLabelColour = "red",
  sigplot.sigLabelSize = 3, sigplot.labelColour = "black",
  sigplot.labelSize = 1, sigplot.labelAngle = 90,
  sigplot.keySize = 1, sigplot.mar = c(5, 4, 4, 2) + 0.1,
  sigplot.width = 7, sigplot.height = 7)
}
\arguments{
\item{object}{Input \code{sig} class object.}

\item{gene_symbol.only}{If to only use probes/genes/genomic features with a gene sybmol id. Default is \code{FALSE}.}

\item{cor.method}{The correlation method, options are "pearson", "spearman" and "pearson". Default is \code{"pearson"}.}

\item{map.colour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n.map.colour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{heatmap.width}{Width for correlation heatmap. Unit is \code{inch}. Default is \code{7}.}

\item{heatmap.height}{Height for correlation heatmap. Unit is \code{inch}. Default is \code{7}.}

\item{sigplot.adj.p}{If to use FDR corrected correlation p value to plot sigPlot. Default is \code{FALSE}.}

\item{sigplot.alpha}{The alpha value for correlation p value. Default is \code{0.05}}

\item{sigplot.sigLabelColour}{The colour for label for the significant pairs. Default is \code{"red"}.}

\item{sigplot.sigLabelSize}{The size for label for the significant pairs. Default is \code{3}.}

\item{sigplot.labelColour}{The significance heatmap axis label colour. Default is \code{"black"}.}

\item{sigplot.labelSize}{The significance heatmap axis label size. Default is \code{1}.}

\item{sigplot.labelAngle}{The significance heatmap axis label angle. Default is \code{90}.}

\item{sigplot.keySize}{The significance heatmap colour key size. Default is \code{1}.}

\item{sigplot.mar}{The A numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot. The default is c(5, 4, 4, 2) + 0.1.}

\item{sigplot.width}{Width for correlation significance plot. Unit is \code{inch}. Default is \code{7}.}

\item{sigplot.height}{Height for correlation significance plot. Unit is \code{inch}. Default is \code{7}.}
}
\value{
A supervised heatmap based on hierarchical clustering analysis in \code{pdf} format, along with correaltion coefficient and p value matrices. The function also outputs a significance plot.
}
\description{
Supervised correlation clustering analysis and heatmap visualization for both microarray and RNAseq, for gene co-expression analysis.
}
\details{
The function takes \code{sig} class object. It behaves in the following ways:
         1. heatmap uses \code{heatmap.2} function from \code{gplots} package.
         2. significance plot uses \code{corrplot} package
         3. the function will use normalized and filtered expression data for both RNAseq and micaorray
         4. the function will automatically subset data using \code{thresholding_summary} from \code{sig} class input
         5. the function will automatically subset and plot for each group under each comparison
}
\examples{
\dontrun{
rbio_supervised_corcluster(object = sig_input, gene_symbol.only = TRUE, cor.method = "pearson",
                           map.colour = "PRGn", n.map.colour = 11,
                           heatmap.width = 7, heatmap.height = 7,
                           sigplot.adj.p = FALSE, sigplot.alpha = 0.05, sigplot.sigLabelColour = "red", sigplot.sigLabelSize = 3,
                           sigplot.labelColour = "black", sigplot.labelSize = 1, sigplot.labelAngle = 90, sigplot.keySize = 1,
                           sigplot.width = 7, sigplot.height = 7)
}
}
