% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{rbio_tom}
\alias{rbio_tom}
\title{rbio_tom}
\usage{
rbio_tom(
  mtx,
  diag = FALSE,
  cor_method = c("pearson", "kendall", "spearman"),
  power = 6,
  tom_type = c("unsigned", "signed"),
  ...,
  manual_membership = FALSE,
  tom_membership = NULL,
  hclust.method = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty",
    "median", "centroid"),
  cutree.method = c("manual", "silhouette", "dynamic"),
  dynamictree.min.size = 20,
  k = NULL,
  h = NULL,
  plot.dendro = TRUE,
  plot.export.name = NULL,
  plot.margins = c(0.5, 0.5, 0.5, 0.5),
  plot.title = NULL,
  plot.title.size = 16,
  plot.dendroline.size = 0.8,
  plot.dendrolabel = TRUE,
  plot.dendrolabel.size = 1.5,
  plot.dendrolabel.space = 2,
  plot.ylabel.size = 1.5,
  plot.width = 150,
  plot.height = 150,
  verbose = TRUE
)
}
\arguments{
\item{mtx}{matrix. Input data matrix. The function calculates TOM distance/affinity for the columns items (or features).}

\item{diag}{Boolean. If to include diagonal for igraph construction. Default is \code{FALSE}.}

\item{cor_method}{String. Correlation method. Default is \code{"pearson"}.}

\item{power}{integer. The power to the correlation coefficients. Default is \code{6}.}

\item{tom_type}{string. Directionality of the TOM analysis. Default is \code{"unsigned"}.}

\item{...}{Additional arguments passed to \code{TOMdist()} function.}

\item{manual_membership}{Boolean. If to use manual membership. Default is \code{FALSE}.}

\item{tom_membership}{Boolean. Set when \code{manual_membership = TRUE}, a \code{named integer vector} containing membership for each features. Default is \code{NULL}.}

\item{hclust.method}{String. Hierarchical clustering distance methods. See the R native \code{\link{hclust}} for details.}

\item{cutree.method}{String. Method to find the optimal k. Default is \code{"silhouette"}.}

\item{dynamictree.min.size}{Integer. When \code{cutree.method = "dynamic"}, the minimum cluster size. Default is \code{20}.}

\item{k}{integer. Set when \code{cutree.method = "manual"}, manually set number of groups to cut to.}

\item{h}{numeric. Set when \code{cutree.method = "manual"}, cut tree height.}

\item{plot.dendro}{Boolean. Whether to plot a dendrogram. Default is \code{TRUE}.}

\item{plot.export.name}{string. The prefix for the exported figure file name. Default is \code{NULL}.}

\item{plot.margins}{numeric vector. Plot margins, unit is "cm", order: \code{t, r, b, l}. Default is \code{c(2, 2, 2, 2)}.}

\item{plot.title}{Boolean. The dendrogram plot title. Default is \code{NULL}.}

\item{plot.title.size}{numeric. The dendrogram plot title size. Default is \code{16}.}

\item{plot.dendroline.size}{numeric. The dendrogram plot line size. Default is \code{1.5}.}

\item{plot.dendrolabel}{Boolean. If to display dendrogram label. Default is \code{TRUE}.}

\item{plot.dendrolabel.size}{numeric. Set when \code{plot.dendrolabel = TRUE}, dendrogram label size. Default is \code{1.5}.}

\item{plot.dendrolabel.space}{numeric. Set when \code{plot.dendrolabel = TRUE}, dendrogram label space. Default is \code{2}.}

\item{plot.ylabel.size}{numeric. Dendrogram y label size. Default is \code{1.5}.}

\item{plot.width}{numeric. The dendrogram plot width. Default is \code{170}.}

\item{plot.height}{numeric. The dendrogram plot height. Default is \code{150}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
An \code{rbio_tom_graph} object with the following items:

        \code{g}: an igraph object for network analysis and visualization.

        \code{tom_distance}: a \code{dist} object for TOM distance.

        \code{tom_affinity}: a \code{matrix} object for TOM affinity.

        \code{tom_dist_hclust}: a \code{hclust} object for TOM distance.

        \code{cutree_method}: method used to cut hclust tree, to generate g membership.

        \code{dynamictree_min_size}: when {cutree.method = "dynamic"}, the minimum tree size for dynamic tree cutting.

        \code{silhouette_score_mean}: when {cutree.method = "silhouette"}, the mean silhouette score.

        \code{k}: when \code{cutree.method = "manual"}, manually set number of groups to cut to.

        \code{h}: when \code{cutree.method = "manual"}, cut tree height.

        \code{manual_membership}: if manually set membership is used.

        \code{tom_membership}: a \code{named integer vector} containing membership for each features clustered.
}
\description{
TOM (topological overlap measure) analysis.
}
\details{
When \code{cutree.method = "manual"}, \code{h} and \code{k} are mutually exclusive. Set one, but not both.

        When {manual _membership = TRUE}, both hclust and tree cutting processes are ignored.
        When {manual _membership = TRUE}, the tom_membership should be a integer vector whose item names are feature names.

        The function uses the \code{"tree"} method from the \code{\link{dynamicTreeCut::cutreeDynamic()}} to cut the tree when
        \code{cutree.method = "dynamic"}.

        The \code{plot.margins} follow the base R setting in \code{\link{par}} for the positioning:
        b: mar[1], l: mar[2], t: mar[3], r: mar[4]
}
\examples{
\dontrun{
           tom_g <- rbio_tom(mtx = t(mtx), k = 10,
                             plot.export.name = "ex1",
                             plot.title = "TOM distance hcluster", plot.title.size = 16,
                             plot.margins = c(0.5 ,0.5, 0.5, 0.5),
                             plot.dendroline.size = 0.2,
                             plot.dendrolabel = TRUE,
                             plot.dendrolabel.size = 1.5,
                             plot.dendrolabel.space = 2,
                             plot.ylabel.size = 1.5)
}
}
