% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_unsupervised_corcluster.rbioarray_de}
\alias{rbio_unsupervised_corcluster.rbioarray_de}
\title{rbio_unsupervised_corcluster.rbioarray_de}
\usage{
\method{rbio_unsupervised_corcluster}{rbioarray_de}(object, ...)
}
\arguments{
\item{object}{Input object in \code{rbioarray_de} class.}

\item{...}{Additional arguments for the default method.}
}
\description{
\code{rbio_unsupervised_corcluster} function for \code{rbioarray_de} class object.
}
