% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_unsupervised_corcluster}
\alias{rbio_unsupervised_corcluster}
\title{rbio_unsupervised_corcluster}
\usage{
rbio_unsupervised_corcluster(object, ...)
}
\arguments{
\item{object}{Input object in either \code{rbioseq_de} or \code{rbioarray_de} class.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
PDF files for unsupervsied correlation heatmap and if set, significance plot. Also CSV files containing unsupervised correlation analysis results.

        It is generally not recommend to perform correlation analysis on the huge number of genes/probes/genomic features - extremely time consuming.
}
\description{
Generic unsupersived correlation clustering function.
}
\details{
It behaves in the following ways:
         1. heatmap uses \code{heatmap.2} function from \code{gplots} package.
         2. significance plot uses \code{corrplot} package

         Since the function depends on the \code{heatmap.2} function from \code{gplots} package for heatmap,
         arguments can be passed directly, as seen in the examples below.
}
\examples{

\dontrun{
# rbioarray_de class input
rbio_unsupervised_corcluster(object = delist, n = 50, cor.method = "pearson",
                             rm.control = TRUE,
                             gene_symbol.only = TRUE,
                             map.colour = "PRGn", n.map.colour = 11,
                             heatmap.axis.label = FALSE,
                             cexCol = 0.6, cexRow = 0.6, margins = c(8, 8), key.title = NA,
                             heatmap.width = 7, heatmap.height = 7,
                             sigplot = TRUE,
                             sigplot.adj.p = FALSE, sigplot.alpha = 0.05,
                             sigplot.sigLabelColour = "red", sigplot.sigLabelSize = 1,
                             sigplot.labelColour = "black", sigplot.labelSize = 0.6, sigplot.labelAngle = 90,
                             sigplot.keySize = 1,
                             sigplot.mar = c(1.25, 1, 1, 0.5),
                             sigplot.width = 7, sigplot.height = 7)

# rbioseq_de class input
rbio_unsupervised_corcluster(object = mrna_de, n = 50, cor.method = "pearson",
                             rm.control = TRUE,
                             gene_symbol.only = TRUE,
                             map.colour = "PRGn", n.map.colour = 11,
                             heatmap.axis.label = FALSE,
                             cexCol = 0.6, cexRow = 0.6, margins = c(8, 8), key.title = NA,
                             heatmap.width = 7, heatmap.height = 7,
                             sigplot = TRUE,
                             sigplot.adj.p = FALSE, sigplot.alpha = 0.05,
                             sigplot.sigLabelColour = "red", sigplot.sigLabelSize = 1,
                             sigplot.labelColour = "black", sigplot.labelSize = 0.6, sigplot.labelAngle = 90,
                             sigplot.keySize = 1,
                             sigplot.mar = c(1.25, 1, 1, 0.5),
                             sigplot.width = 7, sigplot.height = 7)
}

}
