% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbio_unsupervised_corcluster}
\alias{rbio_unsupervised_corcluster}
\alias{rbio_unsupervised_corcluster.rbioarray_de}
\alias{rbio_unsupervised_corcluster.rbioseq_de}
\alias{rbio_unsupervised_corcluster.default}
\title{rbio_unsupervised_corcluster}
\usage{
rbio_unsupervised_corcluster(object, ...)

\method{rbio_unsupervised_corcluster}{rbioarray_de}(object, ...)

\method{rbio_unsupervised_corcluster}{rbioseq_de}(object, sample_id.var.name = NULL, ...)

\method{rbio_unsupervised_corcluster}{default}(
  E,
  genes,
  input.sample_groups,
  n = "all",
  input.comparisons = NULL,
  cor.method = c("pearson", "spearman"),
  rm.control = FALSE,
  input.genes_annotation.control_type,
  gene_symbol.only = FALSE,
  input.genes_annotation.gene_symbol.var_name = NULL,
  input.genes_annotation.gene_id.var_name = NULL,
  export.name = NULL,
  hclust.heatmap = TRUE,
  map.colour = "PRGn",
  n.map.colour = 11,
  ...,
  heatmap.axis.label = FALSE,
  heatmap.width = 7,
  heatmap.height = 7,
  sigplot = TRUE,
  sigplot.adj.p = FALSE,
  sigplot.alpha = 0.05,
  sigplot.sigLabelColour = "red",
  sigplot.sigLabelSize = 3,
  sigplot.labelColour = "black",
  sigplot.labelSize = 1,
  sigplot.labelAngle = 90,
  sigplot.keySize = 1,
  sigplot.mar = c(5, 4, 4, 2) + 0.1,
  sigplot.width = 7,
  sigplot.height = 7,
  verbose = TRUE
)
}
\arguments{
\item{object}{Input object in \code{rbioseq_de} class.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{E}{Expression or count matrix, with rows for genes/probes/genomic features, columns for RNA samples.}

\item{genes}{Annotation data frame for genes/probes/genomic features.}

\item{input.sample_groups}{Input \code{factor} object for sample grouping labels.}

\item{n}{Number of genes/probes/genomic features to be clustered, numeric input or \code{"all"}. Default is \code{"all"}.}

\item{cor.method}{The correlation method, options are "spearman" and "pearson". Default is \code{"pearson"}.}

\item{rm.control}{Whether to remove control probes (Agilent platform) or not. Default is \code{TRUE}.}

\item{input.genes_annotation.control_type}{Only set when \code{rm.control = TRUE}, input control type variable annotation list.}

\item{gene_symbol.only}{Whether or not to remove probes without gene symbol. Default is \code{FALSE}.}

\item{input.genes_annotation.gene_symbol.var_name}{Only set when \code{gene_symbol.only = TRUE}, variable name for gene symbol column in \code{genes} data frame.}

\item{input.genes_annotation.gene_id.var_name}{Only set when \code{gene_symbol.only = TRUE}, variable name for gene id column in \code{genes} data frame.}

\item{export.name}{File name for the export \code{pdf} plot file.}

\item{hclust.heatmap}{Boolean. Whether or not to export a hcluster heatmap. Default is \code{TRUE}.}

\item{map.colour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n.map.colour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{heatmap.axis.label}{Whether to display label for both x- and y- axes. Default is \code{FALSE}.}

\item{heatmap.width}{Width for correlation heatmap. Unit is \code{inch}. Default is \code{7}.}

\item{heatmap.height}{Height for correlation heatmap. Unit is \code{inch}. Default is \code{7}.}

\item{sigplot}{If to show significance plot. Default is \code{TRUE}.}

\item{sigplot.adj.p}{If to use FDR corrected correlation p value to plot sigPlot. Default is \code{FALSE}.}

\item{sigplot.alpha}{The alpha value for correlation p value. Default is \code{0.05}}

\item{sigplot.sigLabelColour}{The colour for label for the significant pairs. Default is \code{"red"}.}

\item{sigplot.sigLabelSize}{The size for label for the significant pairs. Default is \code{3}.}

\item{sigplot.labelColour}{The significance heatmap axis label colour. Default is \code{"black"}.}

\item{sigplot.labelSize}{The significance heatmap axis label size. Default is \code{1}.}

\item{sigplot.labelAngle}{The significance heatmap axis label angle. Default is \code{90}.}

\item{sigplot.keySize}{The significance heatmap colour key size. Default is \code{1}.}

\item{sigplot.mar}{The A numerical vector of the form c(bottom, left, top, right) which gives the number of lines of margin to be specified on the four sides of the plot. The default is c(5, 4, 4, 2) + 0.1.}

\item{sigplot.width}{Width for correlation significance plot. Unit is \code{inch}. Default is \code{7}.}

\item{sigplot.height}{Height for correlation significance plot. Unit is \code{inch}. Default is \code{7}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}
}
\value{
PDF files for unsupervised correlation heatmap and if set, significance plot. Also CSV files containing unsupervised correlation analysis results.
        The function also generates two list objects containing correlation results

        It is generally not recommend to perform correlation analysis on the huge number of genes/probes/genomic features - extremely time consuming.

A heatmap based on hierarchical clustering analysis in \code{pdf} format.
}
\description{
Generic unsupersived correlation clustering function.
}
\details{
It behaves in the following ways:
         1. heatmap uses \code{heatmap.2} function from \code{gplots} package.
         2. significance plot uses \code{corrplot} package

         Since the function depends on the \code{heatmap.2} function from \code{gplots} package for heatmap,
         arguments can be passed directly, as seen in the examples below.

The function uses filtered count data, as opposed to normalized data.
         Due to the compositional nature of NGS data, the count data is transformed using CLR method prior to clustering.
}
\examples{

\dontrun{
# rbioarray_de class input
rbio_unsupervised_corcluster(object = delist, n = 50, cor.method = "pearson",
                             rm.control = TRUE,
                             gene_symbol.only = TRUE,
                             map.colour = "PRGn", n.map.colour = 11,
                             heatmap.axis.label = FALSE,
                             cexCol = 0.6, cexRow = 0.6, margins = c(8, 8), key.title = NA,
                             heatmap.width = 7, heatmap.height = 7,
                             sigplot = TRUE,
                             sigplot.adj.p = FALSE, sigplot.alpha = 0.05,
                             sigplot.sigLabelColour = "red", sigplot.sigLabelSize = 1,
                             sigplot.labelColour = "black", sigplot.labelSize = 0.6, sigplot.labelAngle = 90,
                             sigplot.keySize = 1,
                             sigplot.mar = c(1.25, 1, 1, 0.5),
                             sigplot.width = 7, sigplot.height = 7)

# rbioseq_de class input
rbio_unsupervised_corcluster(object = mrna_de, n = 50, cor.method = "pearson",
                             rm.control = TRUE,
                             gene_symbol.only = TRUE,
                             map.colour = "PRGn", n.map.colour = 11,
                             heatmap.axis.label = FALSE,
                             cexCol = 0.6, cexRow = 0.6, margins = c(8, 8), key.title = NA,
                             heatmap.width = 7, heatmap.height = 7,
                             sigplot = TRUE,
                             sigplot.adj.p = FALSE, sigplot.alpha = 0.05,
                             sigplot.sigLabelColour = "red", sigplot.sigLabelSize = 1,
                             sigplot.labelColour = "black", sigplot.labelSize = 0.6, sigplot.labelAngle = 90,
                             sigplot.keySize = 1,
                             sigplot.mar = c(1.25, 1, 1, 0.5),
                             sigplot.width = 7, sigplot.height = 7)
}

}
