% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rbio_venn_de}
\alias{rbio_venn_de}
\title{rbio_venn_de}
\usage{
rbio_venn_de(
  object,
  gene_symbol.only = FALSE,
  ...,
  plot.Width = 5,
  plot.Height = 5,
  verbose = TRUE
)
}
\arguments{
\item{object}{Input \code{sig} class object.}

\item{gene_symbol.only}{Whether or not to remove probes without gene symbol. Default is \code{FALSE}.}

\item{...}{arguments for \code{vennDiagram()} from \code{limma} package.}

\item{plot.Width}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plot.Height}{The height of the figure for the final output figure file. Default is \code{150}.}
}
\value{
The function outputs a \code{pdf} file for venn diagrams (total, up- and down-regulations). The function also exports overlapping gene or probe into a \code{csv} file.
}
\description{
Venn diagrame for DE results for the \code{sig} class input object.
}
\examples{
\dontrun{
rbio_venn_de(object = siglist, gene_symbol.only = TRUE)
}
}
