% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{rbioarray_corcluster_super}
\alias{rbioarray_corcluster_super}
\title{rbioarray_corcluster_super}
\usage{
rbioarray_corcluster_super(plotName = "data", fltlist = NULL,
  rmControl = TRUE, n_subgroup = NULL, dfmDE = NULL, FDR = TURE,
  q.value = 0.05, FC = NULL, dataProbeVar = NULL, method = "pearson",
  sigPlot = FALSE, cor.sig = 0.05, cor.sigLabelColour = "red",
  cor.sigLabelSize = 3, cor.labelColour = "black", cor.labelSize = 1,
  cor.labelAngle = 90, cor.keySize = 1, axisLabel = FALSE, annot = NULL,
  genesymbolVar = NULL, mapColour = "PRGn", n_mapColour = 11, ...,
  plotWidth = 7, plotHeight = 7)
}
\arguments{
\item{plotName}{File name for the export \code{pdf} plot file. Default is \code{"data"}.}

\item{fltlist}{Based on filtered data, a subset corresponding to the comparasion, either a list, \code{EList} or \code{MAList} object.}

\item{rmControl}{If to remove control probes (Agilent platform). Default is \code{TRUE}.}

\item{n_subgroup}{A vector of sample index (row number) for phenotype group. Default is \code{NULL}. The setting can be obtained from the corresponding condition summary object.}

\item{dfmDE}{A subset of the DE list, i.e. a \code{topTable} dataframe, corresponding to the comparasion (i.e., contrast).}

\item{FDR}{If to use FDR corrcted p value. Default is \code{TRUE}.}

\item{q.value}{P value cut off. Default is \code{0.05}. For unsupervised clustering, set \code{q.value = 1}.}

\item{FC}{Fold change (FC) filter for the heatmap. Default is \code{NULL}.}

\item{dataProbeVar}{\code{dfmDE} variable name for probe name. Default is \code{NULL}.}

\item{method}{The correlation method, options are "pearson", "spearman" and "pearson". Default is \code{"pearson"}.}

\item{sigPlot}{If to generate a significance heatmap. Default is \code{FALSE},}

\item{cor.sig}{Only set when \code{sigPlot = TRUE}, the alpha value for correlation p value. Default is \code{0.05}}

\item{cor.sigLabelColour}{Only set when \code{sigPlot = TRUE}, the colour for label for the significant pairs. Default is \code{"red"}.}

\item{cor.sigLabelSize}{Only set when \code{sigPlot = TRUE}, the size for label for the significant pairs. Default is \code{3}.}

\item{cor.labelColour}{Only set when \code{sigPlot = TRUE}, the significance heatmap axis label colour. Default is \code{"black"}.}

\item{cor.labelSize}{Only set when \code{sigPlot = TRUE}, the significance heatmap axis label size. Default is \code{1}.}

\item{cor.labelAngle}{Only set when \code{sigPlot = TRUE}, the significance heatmap axis label angle. Default is \code{90}.}

\item{cor.keySize}{Only set when \code{sigPlot = TRUE}, the significance heatmap colour key size. Default is \code{1}.}

\item{axisLabel}{Whether to display label for both x- and y- axes. Default is \code{FALSE}.}

\item{annot}{The optional annotation matrix. Only needs to be set to display inforamtions for}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{annot} object. Only set this argument when \code{rowLabel = TRUE}. Default is \code{NULL}.}

\item{mapColour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n_mapColour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{plotWidth}{Width of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{plotHeight}{Height of the plot. Unit is \code{inch}. Default is \code{7}.}
}
\value{
A supervised heatmap based on hierarchical clustering analysis in \code{pdf} format, along with correaltion coefficient and p value matrices. If set, the function also outputs a significant value heatmap.
}
\description{
Wrapper for supervised (or unsupervised) Pearson correlation clustering analysis and heatmap visualization for both microarray and RNAseq, for gene co-expression analysis.
}
\details{
Note that both \code{annot} and \code{genesymbolVar} need to be set to display gene sysmbols as row labels. Otherwise, the row labels will be probe names. Also note that when set to display gene symbols, all the probes without a gene symbol will be removed.
}
\examples{
\dontrun{

# n_subgroup = c(1:4) means the correlation uses samples from 1 to 4 (control in this case).
# The settings can be obtained from the corresponding condition summary object.
rbioarray_corcluster_super(fltlist = all_nrm, n_subgroup = c(1:4),
                           dataProbeVar = "gene_id", FDR = TRUE, q.value = 0.02,
                           dfmDE = all_DE$`conSumPost - conSumPre`,
                           axisLabel = TRUE, genesymbolVar = "gene_name",
                           key.title = "", cexRow = 0.3, cexCol = 0.3, offsetRow = 0.001,
                           offsetCol = 0.001, margins = c(4, 4))

}
}
