% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_DE}
\alias{rbioarray_DE}
\title{rbioarray_DE}
\usage{
rbioarray_DE(objTitle = "data_filtered", fltdata, design, anno,
  multicore = FALSE, ...)
}
\arguments{
\item{objTitle}{Name for the output list. Default is \code{"data_filtered"}.}

\item{fltdata}{filtered data, either a list, \code{EList} or \code{MAList} object.}

\item{design}{Design matrix.}

\item{anno}{Annotation object, usually a \code{dataframe}.}

\item{multicore}{If to use parallel computing. Default is \code{FALSE}.}

\item{...}{arguments for \code{topTable()} from \code{limma} package.}
}
\value{
The function outputs a \code{list} object with DE results, merged with annotation.
}
\description{
DE analysis function.
}
\examples{
\dontrun{
rbiorrary_DE(objTitle = "data", fltdata, design, anno = Anno)
}
}

