% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacyFunctions.R
\name{rbioarray_DE}
\alias{rbioarray_DE}
\title{rbioarray_DE}
\usage{
rbioarray_DE(objTitle = "data_filtered",
  input.outcome.mode = c("categorical", "continuous"),
  output.mode = c("probe.all", "probe.sig", "geneName.all",
  "geneName.sig"), fltlist = NULL, annot = NULL, design = NULL,
  contra = NULL, weights = NULL, ..., plot = TRUE,
  geneName = FALSE, genesymbolVar = NULL, topgeneLabel = FALSE,
  nGeneSymbol = 5, padding = 0.5, FC = 1.5, ctrlProbe = TRUE,
  ctrlTypeVar = "ControlType", sig.method = c("fdr", "spikein",
  "none"), sig.p = 0.05, plotTitle = NULL,
  xLabel = "log2(fold change)", yLabel = "-log10(p value)",
  symbolSize = 2, sigColour = "red", nonsigColour = "gray",
  xTxtSize = 10, yTxtSize = 10, plotWidth = 170, plotHeight = 150,
  parallelComputing = FALSE, clusterType = "PSOCK", verbose = TRUE)
}
\arguments{
\item{objTitle}{Name for the output list. Default is \code{"data_filtered"}.}

\item{input.outcome.mode}{The outcome type. Options are \code{"categorical"} and \code{"continuous"}. Default is \code{"categorical"}.}

\item{output.mode}{Csv file output mode. Options are \code{"probe.all"}, \code{"probe.sig"}, \code{"geneName.all"} and \code{"geneName.sig"}. Default is \code{"probe.all"}. See details below.}

\item{fltlist}{filtered data, either a list, \code{EList} or \code{MAList} object. Default is \code{NULL}.}

\item{annot}{Annotation object, usually a \code{dataframe}. Make sure to name the probe ID variable \code{ProbeName}. Default is \code{NULL}.}

\item{design}{Design matrix. Default is \code{NULL}.}

\item{contra}{Contrast matrix. Default is \code{NULL}.}

\item{weights}{Array weights, determined by \code{arrayWeights()} function from \code{limma} package. Default is \code{NULL}.}

\item{...}{arguments for \code{topTable()} from \code{limma} package.}

\item{plot}{If to generate volcano plots for the DE results. Defualt is \code{TRUE}. Plots are exported as \code{pdf} files.}

\item{geneName}{If to only plot probes with a gene name. Default is \code{FALSE}.}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{annot} object. Only set this argument when \code{geneName = TRUE}. Default is \code{NULL}.}

\item{topgeneLabel}{If to display the gene identification, i.e., probem name or gene name, on the plot. Default is \code{FALSE}.}

\item{nGeneSymbol}{When \code{topgeneLabel = TRUE}, to set how many genes to display. Default is \code{5}.}

\item{padding}{When \code{topgeneLabel = TRUE}, to set the distance between the dot and the gene symbol. Default is \code{0.5}.}

\item{FC}{Threshold for fold change (FC) for volcano plot. Default is \code{1.5}.}

\item{ctrlProbe}{Wether or not the data set has control type variable, with values \code{-1 (negative control)}, \code{0 (gene probes)} and \code{1 (positive control)}. Default is \code{TRUE}.}

\item{ctrlTypeVar}{Set only when \code{ctrlProbe = TRUE}, the control type variable. Default is the \code{Agilent} variable name \code{"ControlType"}.}

\item{sig.method}{DE methods set for p value thresholding. Values are \code{"fdr"}, \code{"spikein"} and \code{"none"}. Default is \code{"fdr"}. \code{"spikein"} can only be set when \code{ctrlProbe = TRUE} and \code{ctrlTypeVar} is properly set.}

\item{sig.p}{Only used when sig.method set as \code{"spikein"}, backup threshold for the p value if spikein p values is larger than \code{0.05}.}

\item{plotTitle}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{symbolSize}{Size of the symbol. Default is \code{2}.}

\item{sigColour}{Colour of the significant genes or probes. Default is \code{"red"}.}

\item{nonsigColour}{Colour of the non-significant genes or probes. Default is \code{"gray"}.}

\item{xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{parallelComputing}{If to use parallel computing. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}

\item{verbose}{Wether to display messages. Default is \code{TRUE}. This will not affect error or warning messeages.}
}
\value{
The function outputs a \code{list} object with DE results, a \code{data frame} object for the F test results, merged with annotation. The function also exports DE reuslts to the working directory in \code{csv} format.
}
\description{
Legacy function. DE analysis function.
}
\details{
When \code{"fdr"} set for sig.method, the p value threshold is set as \code{0.05}.

         When there is no significant genes or probes identified under \code{sig.method = "fdr"}, the threshold is set to \code{sig.p}. When set \code{sig.method = "none"}, the p cutoff will be \code{sig.p}.
         Also note that both \code{geneName} and \code{genesymbolVar} need to be set to display gene sysmbols on the plot. Otherwise, the labels will be probe names. Additionally, when set to display gene symbols,
         all the probes without a gene symbol will be removed.

         The \code{output.mode} argument only affacts output csv files and also not for F test results. This means DE and DE F test resutls for all probes will always be stored to the R environment.
         And F test results for all probes will always be exported as csv file to the working directory.
}
\examples{
\dontrun{
rbioarray_DE(objTitle = "fltdata2", fltlist, annot = Anno, design, contra = contra,
             weights = fltdata$ArrayWeight, parallelComputing = TRUE,
             plot = TRUE, geneName = TRUE, genesymbolVar = "GeneSymbol",
             DE = "spikein")
}
}
