% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_DE}
\alias{rbioarray_DE}
\title{rbioarray_DE}
\usage{
rbioarray_DE(objTitle = "data_filtered", fltdata, anno, design, contra,
  weights = NULL, multicore = FALSE, ..., plot = TRUE, geneName = FALSE,
  genesymbolVar = NULL, signifLabel = FALSE, FC = 1.5, DE = "fdr",
  q.value = 0.05, Title = NULL, xLabel = "log2(fold change)",
  yLabel = "-log10(p value)", symbolSize = 2, xTxtSize = 10,
  yTxtSize = 10, plotWidth = 170, plotHeight = 150)
}
\arguments{
\item{objTitle}{Name for the output list. Default is \code{"data_filtered"}.}

\item{fltdata}{filtered data, either a list, \code{EList} or \code{MAList} object.}

\item{anno}{Annotation object, usually a \code{dataframe}. Make sure to name the probe ID variable \code{ProbeName}}

\item{design}{Design matrix.}

\item{contra}{Contrast matrix.}

\item{weights}{Array weights, determined by \code{arrayWeights()} function from \code{limma} package. Default is \code{NULL}.}

\item{multicore}{If to use parallel computing. Default is \code{FALSE}.}

\item{...}{arguments for \code{topTable()} from \code{limma} package.}

\item{plot}{If to generate volcano plots for the DE results. Defualt is \code{TRUE}. Plots are exported as \code{pdf} files.}

\item{geneName}{If to only plot probes with a gene name. Default is \code{FALSE}.}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{anno} object. Only set this argument when \code{geneName = TRUE}. Default is \code{NULL}.}

\item{signifLabel}{If to display the gene identification, i.e., probem name or gene name, on the plot. Default is \code{FALSE}.}

\item{FC}{Threshold for fold change (FC) for volcano plot. Default is \code{1.5}.}

\item{DE}{DE methods set for p value thresholding. Values are \code{"fdr"} and \code{"spikein"}. Default is \code{"fdr"}.}

\item{q.value}{Only used when DE set as \code{"spikein"}, backup threshold for the p value if spikein p values is larger than \code{0.05}.}

\item{Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{symbolSize}{Size of the symbol. Default is \code{2}.}

\item{xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}
}
\value{
The function outputs a \code{list} object with DE results, merged with annotation. The function also exports DE reuslts to the working directory in \code{csv} format.
}
\description{
DE analysis function.
}
\details{
When \code{"fdr"} set for DE, the p value threshold is set as \code{0.05}. When there is no significant genes or probes identified under \code{DE = "fdr"}, the threshold is set to \code{1}. Also note that both \code{geneName} and \code{genesymbolVar} need to be set to display gene sysmbols as row labels. Otherwise, the lables will be probe names. Additionally, when set to display gene symbols, all the probes without a gene symbol will be removed.
}
\examples{
\dontrun{
tstfunc(objTitle = "fltdata2", fltdata, anno = Anno, design, contra = contra,
        weights = fltdata$ArrayWeight, multicore = TRUE,
        plot = TRUE, geneName = TRUE, genesymbolVar = "GeneSymbol",
        DE = "spikein")
}
}

