% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_filter_combine}
\alias{rbioarray_filter_combine}
\title{rbioarray_filter_combine}
\usage{
rbioarray_filter_combine(
  object,
  filter.percentile = 0.05,
  filter.threshold.min.sample = NULL,
  combine.gene.duplicate = FALSE,
  parallelComputing = FALSE,
  clusterType = "PSOCK",
  verbose = TRUE
)
}
\arguments{
\item{object}{Input \code{rbioarray_plist} object from function \code{\link{rbioarray_transfo_normalize}}.}

\item{filter.percentile}{The percentile threshold for filtering. Default is \code{0.05}. See details for more.}

\item{filter.threshold.min.sample}{Minimum number of samples meeting the filtering threshold. Default is \code{NULL}. See details for more.}

\item{combine.gene.duplicate}{If to combine different transcripts from the same gene/genome feature. Default is \code{FALSE}. See details for more.}

\item{parallelComputing}{Whether to use parallel computing or not. Default is \code{TRUE}.}

\item{verbose}{Whether to display messages. Default is \code{TRUE}. This will not affect error or warning messages.}

\item{cluterType}{clusterType Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
A \code{rbioarray_flist} class, including the following core items:

        \code{E}: filtered (and if set, combined) and normalized expression matrix

        \code{genes}: gene annotation with same row number as E

        \code{gene_duplicates_combined}: either \code{TRUE} or \code{FALSE}

        \code{filter_results}: a list containing \code{neg_control_used}, \code{filter_percentile}, \code{filter_threshold_min_sample}, and \code{filter_summary}

        Additionally, \code{rbioarray_flist} object will include additional items from the input \code{rbioarray_plist} object.
}
\description{
Function to filter, averaging and (if set) combine genes/probes/genome features from the \code{rbioarray_plist} objects.
}
\details{
For \code{filter.percentile}, when the input data has negative probes, the value is set to \code{0.95} so that the 95 percentile of the negative values is the considered the threhold
         When there is no negative control probes, the value is set to \code{0.05}, so that the 5 percentile of the entire data is considered the lower threshold

         For \code{filter.threshold.min.sample}, usually make sure to ensure the target gene has at least three samples, so that stats can be done.

         When \code{combine.gene.duplicate = TRUE},the function combines the expression values from different probes from the same gene.
         This depends on if the input object has a valid gene symbol variable.
}
