% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rbioarray_hcluster_super}
\alias{rbioarray_hcluster_super}
\title{rbioarray_hcluster_super}
\usage{
rbioarray_hcluster_super(plotName = "data", fltDOI, dfmDE, pcutoff = NULL,
  FC = NULL, method = "spikein", fct,
  colGroup = ifelse(length(levels(fct)) < 19, length(levels(fct)), 19),
  distance = "euclidean", clust = "complete", rowLabel = FALSE,
  anno = NULL, genesymbolVar = NULL, colColour = "Paired",
  mapColour = "PRGn", n_mapColour = 11, ..., plotWidth = 7,
  plotHeight = 7)
}
\arguments{
\item{plotName}{File name for the export \code{pdf} plot file. Default is \code{"data"}.}

\item{fltDOI}{Based on filtered data, a subset corresponding to the comparasion, either a list, \code{EList} or \code{MAList} object.}

\item{dfmDE}{A subset of the DE list, i.e. a \code{topTable} dataframe, corresponding to the comparasion (i.e., contrast).}

\item{pcutoff}{P value cut off. Default is \code{NULL}.}

\item{FC}{Fold change (FC) filter for the heatmap. Default is \code{NULL}.}

\item{method}{Thresholding method, "fdr" or "spikein". Default is \code{"spikein"}.}

\item{fct}{Input \code{factor} object for samples.}

\item{colGroup}{Colour group, numeric or dependent on \code{fct}.}

\item{distance}{Distance calculation method. Default is \code{"euclidean"}. See \code{\link{dist}} for more.}

\item{clust}{Clustering method. Default is \code{"complete"}. See \code{\link{hclust}} for more.}

\item{rowLabel}{Whether to display row label or not. Default is \code{FALSE}.}

\item{anno}{Annotation object, usually a \code{dataframe}. Make sure to name the probe ID variable \code{ProbeName}. Only set this argument when \code{rowLabel = TRUE}. Default is \code{NULL}.}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{anno} object. Only set this argument when \code{rowLabel = TRUE}. Default is \code{NULL}.}

\item{colColour}{Column group colour. Default is \code{"Paired"}. See \code{RColorBrewer} package for more.}

\item{mapColour}{Heat map colour. Default is \code{"PRGn"}. See \code{RColorBrewer} package for more.}

\item{n_mapColour}{Number of colours displayed. Default is \code{11}. See \code{RColorBrewer} package for more.}

\item{...}{Additional arguments for \code{heatmap.2} function from \code{gplots} package.}

\item{plotWidth}{Width of the plot. Unit is \code{inch}. Default is \code{7}.}

\item{plotHeight}{Height of the plot. Unit is \code{inch}. Default is \code{7}.}
}
\value{
A supervised heatmap based on hierarchical clustering analysis in \code{pdf} format.
}
\description{
Wrapper for supervised hierarchical clustering analysis and heatmap visualization.
}
\details{
Note that both \code{anno} and \code{genesymbolVar} need to be set to display gene sysmbols as row labels. Otherwise, the row labels will be probe names. Also note that when set to display gene symbols, all the probes without a gene symbol will be removed.
}
\examples{
\dontrun{
rbioarray_hcluster_super(plotName = "pre_experiVnaive", fltDOI = pre_experiVnaive_super, dfmDE = fltdata_DE$pre_experiVnaive,
                         FC = 1.5, pcutoff = 0.0003461,
                         clust = "ward.D2",
                         fct = factor(c("naivepre", "naivepre", "naivepre", "naivepre", "exppre", "exppre", "exppre", "exppre", "exppre"),
                         levels = c("naivepre", "exppre")), trace = "none", srtCol = 30, offsetCol = 0.5, adjCol = c(1, 0),
                         rowLabel = TRUE, anno = Anno, genesymbolVar = "GeneSymbol",
                         offsetRow = 0, adjRow = c(0, 0.5), cexRow = 0.6,
                         key.title = "", keysize = 1.5,
                         key.xlab = "Normalized expression value", key.ylab = "Gene count")
}
}

