% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_rlist}
\alias{rbioarray_rlist}
\title{rbioarray_rlist}
\usage{
rbioarray_rlist(raw.dataframe, raw.annot.var.name = NULL,
  raw.gene_id.var.name = NULL, gene.annot.dataframe = NULL,
  gene.annot.gene_id.var.name = NULL,
  gene.annot.gene_symbol.var.name = NULL, target.annot.file = NULL,
  target.annot.file.path = getwd(), sample_groups.var.name = NULL)
}
\arguments{
\item{raw.dataframe}{Input data frame containing microarray hybridization signals with rows as probe/gene/genomic features and columns as samples. Note: the data frame should contain at least one annotation column.}

\item{raw.annot.var.name}{A string vector containing variable (i.e. column) name(s) for all the annotation columns in \code{raw.dataframe}.}

\item{raw.gene_id.var.name}{Variable (i.e. column) name for gene/probe/genomic feature identification from \code{raw.dataframe}.}

\item{gene.annot.dataframe}{Optional annotation data frame for gene/probe/genomic feature annotation.}

\item{gene.annot.gene_id.var.name}{Set only when \code{gene.annot.dataframe} is provided, variable name for probe/gene/genomic features identification from \code{gene.annot.dataframe}.}

\item{gene.annot.gene_symbol.var.name}{Set only when \code{gene.annot.dataframe} is provided, variable name for probe/gene/genomic features display name from \code{gene.annot.dataframe}, e.g. gene symbols.}

\item{target.annot.file}{File name for the target (i.e. sample) annotation \code{.csv} file.}

\item{target.annot.file.path}{The directory for \code{target.annot.file}. Default is \code{getwd()}.}

\item{sample_groups.var.name}{The variable name for sample groupping information from \code{target.annot.file}.}
}
\value{
A \code{rbioarray_rlist} object. The \code{rbioarray_rlist} class contains the following items:

        \code{E}: raw expression (i.e. hybridization signal)

        \code{raw_file.gene_annotation.var_name}

        \code{raw_file.gene_id.var_name}

        \code{gene_display_name_used}: if the gene display name is extracted from \code{gene.annot.dataframe}.

        \code{genes_annotation.gene_id.var_name}

        \code{genes_annotation.gene_symbol.var_name}

        \code{targets}: the sample annotation data frame.

        \code{sample_groups}
}
\description{
Function to constuct \code{rbioarray_rlist} class object from microarary and annotation data. The \code{rbioarray_rlist} object is the starting point for all microarray data analysis.
}
\details{
The word "gene" used in argument names and output item names is in its broader meaning of gene/probe/genomic feature.

         The optional annotation data frame \code{gene.annot.dataframe} should contain any additional annotation information in addition to the annotation column(s) from \code{raw.dataframe}.
         It is noted that \code{gene.annot.dataframe} should contain at least one column for the sample type of gene/probe/genomic feature identification as the identification variable from  \code{raw.dataframe}.
         Such column is set via argument \code{raw_file.gene_id.var_name} and \code{genes_annotation.gene_id.var_name}.
         The gene display name will only be used if \code{gene.annot.dataframe} is used. Otherwise, it wil use \code{raw.gene_id.var.name} from \code{raw.dataframe}.

         To keep things consistent with the \code{Elist} from the dependent \code{limma} package. The \code{rbioarray_rlist} class contains many common elements from \code{Elist} class.
}
\examples{
\dontrun{
raw_list <- rbioarray_rlist(raw.dataframe = raw, raw.annot.var.name = c("PROBE_ID","ILMN_GENE"), raw.gene_id.var.name = "PROBE_ID",
                            gene.annot.dataframe = annot, gene.annot.gene_id.var.name = "PROBE_ID", gene.annot.gene_symbol.var.name = "SYMBOL",
                            target.annot.file = "sample_index_end6.csv", sample_groups.var.name = "time_point")
}
}
