% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayDEA.R
\name{rbioarray_transfo_normalize}
\alias{rbioarray_transfo_normalize}
\title{rbioarray_transfo_normalize}
\usage{
rbioarray_transfo_normalize(object, ...)
}
\arguments{
\item{object}{Input obejct with raw data and annotation information. Should be a \code{rbioarray_rlist} class.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
A \code{rbioarray_plist} class object with the following items:

        \code{E}: Normalized and then log2 transformed expression matrix

        \code{design}: Microrray experiment sample design matrix

        \code{ArrayWeight}

        \code{original_E}: Original E matrix from the input \code{rbioarray_rlist}.

        Additionally, \code{rbioarray_plist} object will include additional items from the input \code{rbioarray_rlist} object.
}
\description{
Generic data log transformation and nomalization function for microarray.
}
\details{
The \code{rbioarray_rlist} object can be obtained from \code{\link{rbioarray_rlist}} function.

         The expression matrix will be normalized and then log2 tranformed for output.

         A note to the \code{normalizeBetweenArrays} function from \code{limma} package:
         The function normalizes data BEFORE log2 transformation when the input is \code{EListRaw} object.
         However, when input is \code{matrix}, it assumes the data has already been log2 tranformed, meaning normalization will
         be done AFTER log2 transformation.
            After comparision, these two methods will NOT produce the same results for the same expression data:
         In other words, applying \code{normalizeBetweenArrays} directly to log2 transformed E matrix is NOT the same as apply the function
         to the \code{EListRaw} that contains E.
            In fact, when using \code{"quantile"} method, applying \code{normalizeBetweenArrays} to log2 transformed E matrix is the
         same as applying \code{normalizeBetweenArrays} to the \code{EListRaw} that has object$E using \code{"cyclicloess"} method.
         Indeed, the source code of \code{normalizeBetweenArrays} suggests that's the case since log2 transformation is appled BEFORE
         \code{"cyclicloess"} normalization. However, transformation happens AFTER normalization for \code{"quantile"} and other methods.
}
\examples{

\dontrun{
# for \\code{rbioarray_rlist} object:
microarray_plist <- rbioarray_transfo_normalize(object = microarray_rlist, logTransfo = TRUE, logTransfo.method = "log2",
                                                logTransfo.parallelComputing = TRUE, logTransfo.cluterType = "FORK",
                                                bgc.method = "auto", between.sample.norm.method = "quantile")
}

}
