% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS}
\alias{rbioGS}
\title{rbioGS}
\usage{
rbioGS(GS, pVar, logFCVar, tVar, idVar, method_p = c("fisher", "stouffer",
  "reporter", "tailStrength", "wilcoxon"), method_t = c("page", "gsea",
  "maxmean"), ..., parallelComputing = FALSE, clusterType = "PSOCK")
}
\arguments{
\item{GS}{Pre-loaded gene set objects.}

\item{pVar}{Gene level p values. Could be, but not exclusive to, a variable of a dataframe. Must be the same length as \code{logFCVar}, \code{tVar} and \code{idVar}.}

\item{logFCVar}{Gene level logFC (log fold change). Could be, but not exclusive to, a variable of a dataframe. Must be the same length as \code{pVar}, \code{tVar} and \code{idVar}.}

\item{tVar}{Gene leve t values. Could be, but not exclusive to, a variable of a dataframe. Must be the same length as \code{pVar}, \code{logFCVar} and \code{idVar}.}

\item{idVar}{Gene IDs. Could be, but not exclusive to, a variable of a dataframe. Must be the same length as \code{pVar}, \code{logFCVar} and \code{tVar}. Currently only takes \code{Entrez ID}.}

\item{method_p}{Gene set ernichment methods that takes \code{p value} and \code{logFC}. Default is \code{c("fisher", "stouffer", "reporter", "tailStrength", "wilcoxon")}.}

\item{method_t}{Gene set ernichment methods that takes \code{t statistics}. Default is \code{c("page", "gsea", "maxmean")}.}

\item{...}{Arguments to pass to \code{runGSA} function from \code{piano} pacakge. See the corresponding help page from of \code{piano} for details.}

\item{parallelComputing}{If to use parallel computing or not. Default is \code{FALSE}}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs a \code{list} object with GSA results, both for p value based and t value based.
}
\description{
Add Human entrez ID to the DE dataframe
}
\details{
The function is based on piano package. It runs "fisher", "stouffer", "reporter", "tailStrength", "wilcoxon" for p value based GSA, and "page", "gsea", "maxmean" for t value based GSA.
}
\examples{
\dontrun{
gsoutput <- rbioGS(GS = kegg, pVar = dfm$p_value, logFCVar = dfm$logFC, tVar = dfm$t_value, idVar = dfm$EntrezID, nPerm = 1000, parallelComputing = TRUE, clusterType = "FORK")


}
}
