% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_all}
\alias{rbioGS_all}
\title{rbioGS_all}
\usage{
rbioGS_all(fileName, input, entrezVar = NULL, GS = NULL, plot = TRUE,
  multicore = FALSE, clusterType = "PSOCK")
}
\arguments{
\item{fileName}{Output figure file name.}

\item{input}{Input list object for DE results.}

\item{entrezVar}{Name of the EntrezID variable in the \code{input} object.}

\item{GS}{Pre-loaded gene set objects.}

\item{multicore}{If to use parallel computing or not. Default is \code{FALSE}}

\item{clusterType}{Only set when \code{multicore = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
Outputs  \code{csv} files and \code{pdf} figure files with GSA results.
}
\description{
Add Human entrez ID to the DE dataframe
}
\details{
This is an all-in-one function for GS anlyasis based on piano package. It runs "fisher", "stouffer", "reporter", "tailStrength", "wilcoxon" for p value based GSA, and "page", "gsea", "maxmean" for t value based GSA.
}
\examples{
\dontrun{
rbioGS_all(GS = kegg, pVar = dfm$p_value, logFCVar = dfm$logFC, tVar = dfm$t_value, idVar = dfm$EntrezID, multicore = TRUE, clusterType = "FORK")


}
}

