% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_boxplot}
\alias{rbioGS_boxplot}
\title{rbioGS_boxplot}
\usage{
rbioGS_boxplot(GS_list, ..., GS = "OTHER", fileName = "GS_list",
  plotTitle = NULL, xLabel = "rank", yLabel = NULL, plotWidth = 170,
  plotHeight = 150)
}
\arguments{
\item{GS_list}{piano GS results object.}

\item{...}{Arguments passing to \code{consensusScores} function from \code{piano} package. See the corresponding \code{piano} package help page for details.}

\item{GS}{if GS = \code{"KEGG"}, the function will remove the "KEGG_" string in the GS name variable. Default is \code{"OTHER"}.}

\item{fileName}{Output file name. Default is \code{"GS_list"}.}

\item{plotTitle}{Title of the plot. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Default is \code{"rank"}.}

\item{yLabel}{Y-axis label. Default is \code{NULL}.}

\item{plotWidth}{Set the width of the plot. Default is \code{170}.}

\item{plotHeight}{Set the height of the plot. Default is \code{150}.}
}
\value{
Outputs a \code{pdf} boxplot figure file with allGSA results.
}
\description{
Generate boxplot from piano GS rank object,
}
\details{
The function is a wrapper that takes resulted object from \code{runGSA} function from \code{piano} package.
}
\examples{
\dontrun{

rbioGS_boxplot(GS_object, fileName = "GS_analysis", class = "mixed", direction="down", n = 15, adjusted = TRUE, method = "median", rowNames = "names", plotWidth = 260, plotHeight = 240)

}
}

