% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_scatter}
\alias{rbioGS_scatter}
\title{rbioGS_scatter}
\usage{
rbioGS_scatter(GSAList, ..., rankCutoff, pCutoff, fileName, plotWidth = 170,
  plotHeight = 150)
}
\arguments{
\item{...}{Arguments passing to \code{consensusHeatmap} function from \code{piano} package. See the responding help page of \code{piano} for details.}

\item{rankCutoff}{Cutoff value for GS rank line.}

\item{pCutoff}{Cutoff value for GS p value line.}

\item{fileName}{Output file name.}

\item{plotHeight}{Set the height of the plot. Default is \code{150}.}

\item{GS_list}{GSA list generated from \code{\link{rbioArray_allGSA}}.}
}
\value{
Outputs a \code{pdf} scatter figure file with allGSA results.
}
\description{
Generate scatter plot for piano GS rank heatmap obejct.
}
\details{
The function is based on piano package.
}
\examples{
\dontrun{

rbioGS_scatter(GS_Pos, cutoff = 15, method = "median", adjusted = TRUE, rankCutoff = 50, pCutoff = 0.05, fileName = "GS_pos")

}
}

