% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSA.R
\name{rbioGS_sp2hsaEntrez}
\alias{rbioGS_sp2hsaEntrez}
\title{rbioGS_sp2hsaEntrez}
\usage{
rbioGS_sp2hsaEntrez(DElst, tgtSpecies = "mmu", ensemblTransVar = NULL,
  parallelComputing = FALSE)
}
\arguments{
\item{DElst}{The list with DE reuslt, from functions \code{\link{rbioarray_DE}} or \code{\link{rbioseq_DE}}.}

\item{tgtSpecies}{The target species. Options are \code{"mmu"} and \code{"rno"}.}

\item{ensemblTransVar}{The name of the variable from DE list containing ensembl transcript ID.}

\item{parallelComputing}{If to use parallel computing. The cluster mode is \code{PSOCK} for now. Default is \code{FALSE}.}
}
\value{
Outputs a DE \code{list} object with human Entrez ID for each dataframe. This list has the exact same format as the input DE list.
}
\description{
convert from mouse/rat ensemble transcript ID to and add Human entrez ID to the DE list from the DE functions, i.e. \code{\link{rbioarray_DE}} or \code{\link{rbioseq_DE}}, as human EntrezID is needed for GS analysis if using human gene sets.
}
\examples{
\dontrun{
rbioGS_sp2hsaEntrez(DElst = comparison_DE, tgtSpecies = "mmu", ensemlTransVar = "EnsemblID")
}
}
