% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{rbioSeq_DE}
\alias{rbioSeq_DE}
\title{rbioSeq_DE}
\usage{
rbioSeq_DE(AnnCountDfm, DEGNum = Inf, adjMethod = "fdr")
}
\arguments{
\item{AnnCountDfm}{Input data frame with annoatation merged.}

\item{DEGNum}{Number of targets to display. Default is \code{Inf}.}

\item{adjMethod}{P value correct methods. values: \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, \code{"none"}.}
}
\value{
Outputs a \code{dataframe} object DE resutls.
}
\description{
The function below is an all-in-one solution to get DGE list and itrequires limma and edgeR packages for RNA-seq dataset.
}
\examples{
\dontrun{
DE_dataframe <- rbioArray_DE(dataframe)
}
}

