% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacyFunctions.R
\name{rbioseq_DE}
\alias{rbioseq_DE}
\title{rbioseq_DE}
\usage{
rbioseq_DE(
  objTitle = "data_filtered",
  dfm_count = NULL,
  dfm_annot = NULL,
  count_threshold = "none",
  norm.method = "TMM",
  qc_plot = TRUE,
  design = NULL,
  contra = NULL,
  ...,
  plot = TRUE,
  geneName = FALSE,
  genesymbolVar = NULL,
  topgeneLabel = FALSE,
  nGeneSymbol = 5,
  padding = 0.5,
  FC = 1.5,
  FDR = TRUE,
  sig.p = 0.05,
  Title = NULL,
  xLabel = "log2(fold change)",
  yLabel = "-log10(p value)",
  symbolSize = 2,
  sigColour = "red",
  nonsigColour = "gray",
  xTxtSize = 10,
  yTxtSize = 10,
  plotWidth = 170,
  plotHeight = 150,
  parallelComputing = FALSE,
  clusterType = "PSOCK"
)
}
\arguments{
\item{objTitle}{Name for the output list. Default is \code{"seq_data"}.}

\item{dfm_count}{Dataframe contains the feature read counts, with rows as genomic featues (or genes) and column as samples. Default is \code{NULL}.}

\item{dfm_annot}{Dataframe contains the gene annotation information, with rows as genmic features and columns as annotation variables. The row lengths of this dataframe should be the same as \code{dfm_count}.}

\item{count_threshold}{Read count threshold. No filtering will be applied when set \code{"none"}. Otherwise, a numeric number can be set as the minimum read count for filtering. Default is \code{"none"}.}

\item{norm.method}{Normalization methods.Currently, the none-compositional methods are supported. Options are \code{"none", "TMM","RLE","upperquartile"}. Default is \code{"TMM"}. See \code{calcNormFactors} function from  \code{edgeR} pacakge for more.}

\item{qc_plot}{Wether or not to produce a QC plot upon filtering, normalization and weight calculation. Default is \code{TRUE}.}

\item{design}{Design matrix. Default is \code{NULL}.}

\item{contra}{Contrast matrix. Default is \code{NULL}.}

\item{...}{arguments for \code{topTable()} from \code{limma} package.}

\item{plot}{If to generate volcano plots for the DE results. Defualt is \code{TRUE}. Plots are exported as \code{pdf} files.}

\item{geneName}{If to only plot probes with a gene name. Default is \code{FALSE}.}

\item{genesymbolVar}{The name of the variable for gene symbols from the \code{annot} object. Only set this argument when \code{geneName = TRUE}. Default is \code{NULL}.}

\item{topgeneLabel}{If to display the gene identification, i.e., probem name or gene name, on the plot. Default is \code{FALSE}.}

\item{nGeneSymbol}{When \code{topgeneLabel = TRUE}, to set how many genes to display. Default is \code{5}.}

\item{padding}{When \code{topgeneLabel = TRUE}, to set the distance between the dot and the gene symbol. Default is \code{0.5}.}

\item{FC}{Threshold for fold change (FC) for volcano plot. Default is \code{1.5}.}

\item{FDR}{Wether or not using FDR p value correction. Default is \code{TRUE}.}

\item{sig.p}{Threshold for the p value. Default is \code{0.05}.}

\item{Title}{Figure title. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{xLabel}{X-axis label. Make sure to use quotation marks. Use \code{NULL} to hide. Default is \code{NULL}.}

\item{yLabel}{Y-axis label. Make sure to use quotatio marks. Use \code{NULL} to hide. Default is \code{"Mean Decrease in Accurac"}}

\item{symbolSize}{Size of the symbol. Default is \code{2}.}

\item{sigColour}{Colour of the significant genes or probes. Default is \code{"red"}.}

\item{nonsigColour}{Colour of the non-significant genes or probes. Default is \code{"gray"}.}

\item{xTxtSize}{Font size for the x-axis text. Default is \code{10}.}

\item{yTxtSize}{Font size for the y-axis text. Default is \code{10}.}

\item{plotWidth}{The width of the figure for the final output figure file. Default is \code{170}.}

\item{plotHeight}{The height of the figure for the final output figure file. Default is \code{150}.}

\item{parallelComputing}{If to use parallel computing. Default is \code{FALSE}.}

\item{clusterType}{Only set when \code{parallelComputing = TRUE}, the type for parallel cluster. Options are \code{"PSOCK"} (all operating systems) and \code{"FORK"} (macOS and Unix-like system only). Default is \code{"PSOCK"}.}
}
\value{
The function outputs a \code{dataframe} object with filtered and normalized readcounts, a \code{list} object with DE results, a \code{dataframe} object for the F test results, merged with annotation. The function also exports DE reuslts to the working directory in \code{csv} format.
}
\description{
Legacy function. DE analysis function for RNA-seq data, with count filtering functionality.
}
\details{
Sample weights are automatically caluclated during Voom normalization. When \code{count_threshold = 0}, the function will filter out all the genomic features with a total read count 0. When there is no significant genes or probes identified under \code{FDR = TRUE}, the function conducts DE analysis without p value correction and outputs an warning. Also note that both \code{geneName} and \code{genesymbolVar} need to be set to display gene sysmbols on the plot. Additionally, when set to display gene symbols, all the features without a gene symbol will be removed.
}
\examples{
\dontrun{
rbioseq_DE(objTitle = "test", dfm_count = Ann_Count_all[, 5:12], dfm_annot = Ann_Count_all[, 1:4], count_threshold = 50,
           design = design, contra = contra,
           FDR = TRUE, sig.p = 0.05,
           geneName = TRUE, genesymbolVar = "gene_name", topgeneLabel = TRUE, nGeneSymbol = 10)
}
}
