% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqDEA.R
\name{rbioseq_import_gtf}
\alias{rbioseq_import_gtf}
\title{Title rbioseq_import_gtf}
\usage{
rbioseq_import_gtf(file)
}
\arguments{
\item{file}{GTF/GFF file. Add path if needed.}
}
\value{
A matrix with items from GTF/GFF file.
}
\description{
Import GTF/GFF files
}
\details{
The following items are extracted from GTF/GFF file: \code{chromosome}, \code{gene_id}, \code{gene_type}, and \code{gene_name}.
}
\examples{

\dontrun{
gtf <- rbioseq_import_gtf(file = "gencode_v19.gtf")
}

}
