% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_count_to_de.R
\name{rnaseq_de.default}
\alias{rnaseq_de.default}
\title{Title rnaseq_de.default}
\usage{
\method{rnaseq_de}{default}(x, y = NULL, y.gene_id.var.name = "genes",
  y.gene_symbol.var.name = "genes", filter.threshold.cpm = "none",
  filter.threshold.min.sample = NULL, annot.group = NULL,
  between.genes.norm.method = "TMM", design, contra, qc.plot = TRUE)
}
\arguments{
\item{x}{Input read count matrix, with rows for genes and columns for RNA samples}

\item{y}{Target (e.g. genes) annotation matrix or vector.}

\item{y.gene_id.var.name}{Variable name for gene (i.e. target) identification. Default is \code{"genes"}.}

\item{y.gene_symbol.var.name}{Variable name for gene (i.e. target) symbols. Default is \code{"genes"}.}

\item{filter.threshold.cpm}{Filtering threshold for counts based on CPM (counts per million). Default is \code{"none"}.}

\item{filter.threshold.min.sample}{Minimum number of samples meeting the count threshold. Default is \code{NULL}.}

\item{annot.group}{Set only when \code{filter.threshold.min.sample} is not \code{NULL}, sample group annotation object. Can be a \code{factor} or \code{vector} object.}

\item{between.genes.norm.method}{Between gene normalization method. Options are: \code{"none"}, \code{"TMM"}, \code{"RLE"}, \code{"upperquartile"}. Default is \code{"TMM"}.}

\item{design}{Design matrix.}

\item{contra}{Contrast matrix.}

\item{qc.plot}{QC plot for the input read counts}
}
\value{
A \code{rbioseq_de} object. The items of the object are following:

\code{filter_results}

\code{normalization_results}

\code{gene_id_var_name}

\code{gene_symbol_var_name}

\code{F_stats}

\code{DE_results}
}
\description{
The default \code{rnaseq_de} function
}
\details{
\code{filter.threshold.cpm.count} uses CPM (counts per million) as the basis for filtering. The rule of thumb is to filter reads independently from the groupping information.

The default is based on the paper by Chen et al (2016):
10~15 counts per library, which translate to 10~15/L with L being the smallest library size in millions.
Therefore, if using 10 as an example, the above becomes 10 / min(library_sizes) / 1000000

Chen W, Lun ATL, Smyth GK. 2016. From reads to genes to pathways: differential expression analysis of RNA-Seq experiments
using Rsubread and the edgeR quasi-likelihood pipeline. F1000Research. 5:1438.
}
