% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_count_to_de.R
\name{rnaseq_de}
\alias{rnaseq_de}
\title{Title rnaseq_de}
\usage{
rnaseq_de(object, ...)
}
\arguments{
\item{object}{Object containing expression values. Currently the function supports \code{rbioseq_count} and \code{mir_count} from \code{RBioMIR} pacakge.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
A differential expression result object.
}
\description{
Generic RNAseq differential expression analysis function
}
\examples{

\dontrun{
rnaseq_de(object = mrnaseq)
}

}
