% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test.R
\name{sig.rbioseq_de}
\alias{sig.rbioseq_de}
\title{Title sig.rbioseq_de}
\usage{
\method{sig}{rbioseq_de}(object, export.name = NULL)
}
\arguments{
\item{object}{A \code{rbioseq_de} object \code{\link{rnaseq_de}} function.}

\item{export.name}{Optional name used for output objects to the environment and directory. Default is \code{NULL}.}

\item{...}{Additional arguments for \code{\link{sig.defuault}}.}
}
\description{
The \code{sig} function for \code{rbioseq_de} object from \code{\link{rnaseq_de}} function.
}
