% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_test.R
\name{sig}
\alias{sig}
\title{Title sig}
\usage{
sig(object, ...)
}
\arguments{
\item{object}{Object containing DE inforamtion. Currently the function supports \code{rbioseq_de} and \code{rbioarray_de} objects.}

\item{...}{Additional arguments for corresponding S3 class methods.}
}
\value{
A differential expression result object.
}
\description{
Generic significance test function
}
\examples{

\dontrun{
sig(object = mrnade)
}

}
